% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_importance.R
\name{aggregate_importance}
\alias{aggregate_importance}
\title{Aggregates importance data frame with spatial predictors}
\usage{
aggregate_importance(x)
}
\arguments{
\item{x}{An importance data frame with spatial predictors.}
}
\value{
A data frame with aggregated importance scores for the spatial predictors.
}
\description{
Aggregates the importance of spatial predictors within an importance data frame into min, max, mean, and median.
}
\examples{
\dontrun{
if(interactive()){
 data(distance_matrix)
 data(plant_richness_df)

 model <- rf_spatial(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds =  c(0, 500, 1000)
 )

 aggregate_importance(model$variable.importance$per.variable)

 }
}
}
