% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area under the ROC curve}
\usage{
auc(o, p)
}
\arguments{
\item{o}{Numeric vector with observations, must have the same length as \code{p}.}

\item{p}{Numeric vector with predictions, must have the same length as \code{o}.}
}
\value{
Numeric, AUC value.
}
\description{
Computes the area under the ROC curve in models with binary responses.
}
\examples{
\donttest{
if(interactive()){

 out <- auc(
   o = runif(10),
   p = runif(10)
   )

}
}
}
