% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_vif.R
\name{auto_vif}
\alias{auto_vif}
\alias{vif}
\title{Multicollinearity reduction via Variance Inflation Factor}
\usage{
auto_vif(
  data = NULL,
  predictor.variable.names = NULL,
  preference.order = NULL,
  vif.threshold = 5,
  verbose = TRUE
)

vif(data)
}
\arguments{
\item{data}{A data frame with predictors, or alternatively, the result of \code{\link[=auto_cor]{auto_cor()}}. Default: \code{NULL}.}

\item{predictor.variable.names}{Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data are used, and non-numeric ones discarded. Default: \code{NULL}}

\item{preference.order}{Character vector indicating the user's order of preference to keep variables. Predictors not included in this argument are ranked by their VIF. Default: \code{NULL}.}

\item{vif.threshold}{Numeric between 0 and 10 defining the selection threshold for the VIF analysis. Higher numbers result in a more relaxed variable selection. Default: 5.}

\item{verbose}{Logical. if \code{TRUE}, describes the function operations to the user. Default:: \code{TRUE}}
}
\value{
List with three slots:
\itemize{
\item \code{vif}: data frame with the names of the selected variables and their respective VIF scores.
\item \code{selected.variables}: character vector with the names of the selected variables.
\item \code{selected.variables.df}: data frame with the selected variables.
}
}
\description{
Selects predictors that are not linear combinations of other predictors by using computing their variance inflation factors (VIF). Allows the user to define an order of preference for the selection of predictors. \strong{Warning}: variables in \code{preference.order} not in \code{colnames(x)}, and non-numeric columns are removed from \code{x} and \code{preference.order}. The same happens with rows having NA values (\code{\link[=na.omit]{na.omit()}} is applied). The function issues a message if zero-variance columns are found. Notice that identical columns with different names may crash this function, but you can prevent this by running \code{\link[=auto_cor]{auto_cor()}} before \code{\link[=auto_vif]{auto_vif()}}.
}
\details{
This function has two modes of operation:
\itemize{
\item 1. When the argument \code{preference.order} is \code{NULL}, the function removes on each iteration the variable with the highest VIF until all VIF values are lower than \code{vif.threshold}.
\item 2. When \code{preference.order} is provided, the variables are selected by giving them priority according to their order in \code{preference.order}. If there are variables not in \code{preference.order}, these are selected as in option 1. Once both groups of variables have been processed, all variables are put together and selected by giving priority to the ones in \code{preference.order}. This method preserves the variables desired by the user as much as possible.
}
Can be chained together with \code{\link[=auto_cor]{auto_cor()}} through pipes, but it is always recommended to run \code{\link[=auto_cor]{auto_cor()}} first, since pairs of variables with a Pearson correlation index of 1 may crash \code{\link[=auto_vif]{auto_vif()}}.
}
\examples{
if(interactive()){

data(
  ecoregions_df,
  ecoregions_predictor_variable_names
  )

#on a data frame
out <- auto_vif(
 data = ecoregions_df,
 predictor.variable.names = ecoregions_predictor_variable_names
 )

#getting out the vif data frame
out$vif

#getting the names of the selected variables
out$selected.variables

#getting the data frame of selected variables
out$selected.variables.df

#with preference order
 out <- auto_vif(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names,
   preference.order = ecoregions_predictor_variable_names[1:5],
 )

#with pipes
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names
 ) \%>\%
 auto_vif()

 #vif function
 vif(data = ecoregions_df[, ecoregions_predictor_variable_names])


}
}
\seealso{
\code{\link[=auto_cor]{auto_cor()}}
}
