% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{case_weights}
\alias{case_weights}
\title{Generates case weights for binary responses}
\usage{
case_weights(x = NULL, case.weights = NULL)
}
\arguments{
\item{x}{Numeric vector with values 1 and 0 representing the response column to train a binary-response model. Default: \code{NULL}}

\item{case.weights}{Numeric vector with case weights. Only for internal use within the package. Default: \code{NULL}}
}
\value{
A vector with a length equal to \code{x} with case weights.
}
\description{
When the data is binary, setting the \code{ranager} argument \code{case.weights} helps to minimize the issues produced by class imbalance. This function takes a binary response variable and returns a vector of weights populated with the values \verb{1/#zeros} and \verb{1/#ones}. It is used internally by the function \code{\link[=rf]{rf()}}.
}
\examples{
if(interactive()){

 case_weights(
   x = c(0, 0, 0, 1, 1)
 )

 }
}
