% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_specification.R
\name{cluster_specification}
\alias{cluster_specification}
\title{cluster_specification}
\usage{
cluster_specification(ips = NULL, cores = NULL, user = Sys.info()[["user"]])
}
\arguments{
\item{ips}{character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cores}{numeric integer vector, number of cores on each machine.}

\item{user}{character string, name of the user (should be the same throughout machines), Defaults to the current system user.}
}
\value{
a list ready to be used as input for the "spec" argument of the function \link[parallel]{makeCluster}.
}
\description{
specifies a cluster for \link[parallel]{makeCluster}
}
\details{
This function is used internally by several other functions of the package, but can be useful for a user using \link[foreach]{foreach} to parallelize loops in a cluster.
}
\examples{
 cluster.spec <- cluster_specification(
   ips = c("10.0.1.40", "10.0.1.41"),
   cores = c(6, 4),
   user = Sys.info()[["user"]]
 )
 cluster.spec
}
