% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\docType{data}
\name{distance_matrix}
\alias{distance_matrix}
\title{Matrix of distances among the World's Ecoregions.}
\format{
A matrix with 662 rows and columns.
}
\usage{
distance_matrix
}
\description{
Distance matrix in km among the ecoregions described in the \link{plant_richness} dataset.
distance_matrix
}
\keyword{datasets}
