% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_center_distance_matrix.R
\name{double_center_distance_matrix}
\alias{double_center_distance_matrix}
\title{double_center_distance_matrix}
\usage{
double_center_distance_matrix(x = NULL, distance.threshold = 0)
}
\arguments{
\item{x}{distance matrix among the records represented in the numeric vector. The number of rows of this matrix must be equal to the length of x. Default: NULL, Default: NULL}

\item{distance.threshold}{(optional) numeric, positive, in the range of values of \code{distance.matrix} distances below this value in the distance matrix are set to 0.  Default: 0}
}
\value{
A double-centered matrix of the same dimensions as \code{x}.
}
\description{
generates a double-centered matrix from the weights of a distance matrix (see \link{weights_from_distance_matrix}) and a distance threshold. This is a required step before the computation of Moran's Eigenvector Maps.
}
\details{
Applies \link{weights_from_distance_matrix} to the distance matrix \code{x},

ach element has subtracted fromit its corresponding row mean and column mean, and we add back the averageof all the elements.. Based on \link[ade4]{bicenter.wt} by Daniel Chessel.
}
\examples{
\dontrun{
if(interactive()){
 data(distance_matrix)
 x <- double_center_distance_matrix(
   x = distance_matrix
 )
 x
 }
}
}
