% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_continuous_response}
\alias{ecoregions_continuous_response}
\title{Name of the continuous dependent variable}
\format{
Character string.
}
\usage{
data(ecoregions_continuous_response)
}
\description{
Character string with the name of the dependent variable.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
