% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_df.R
\docType{data}
\name{ecoregions_depvar_name}
\alias{ecoregions_depvar_name}
\title{Name of the dependent variable}
\format{
Character string.
}
\usage{
data(ecoregions_depvar_name)
}
\description{
Character string with the name of the dependent variable.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
