% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_df}
\alias{ecoregions_df}
\title{Plant richness and predictors of American ecoregions}
\format{
A data frame with 225 rows and 49 columns:
\itemize{
\item \code{ecoregion_id}: Numeric id of the ecoregion.
\item \code{ecoregion_name}: Name of the ecoregion according to \href{https://ecoregions2017.appspot.com/}{Ecoregions 2017}.
\item \code{x} and \code{y}: Numeric, coordinates of the ecoregion centroids, in EPSG 4326.
\item \code{plant_richness}: Number of vascular species found in the ecoregion. Response variable.
\item \code{plant_richness_binary}: Binary version of \code{plant_richness}, with value \code{1} when \code{plant_richness} is higher than 5000, and \code{0} otherwise.
\item \code{ecoregion_area_km2}: Area of the ecoregion in squared kilometers.
\item \code{sampling_bias}: Indicator of sampling bias, measured as number of vascular plant species found in the ecoregion divided by the number of GBIF records of the ecoregion, divided by the area of the ecoregion.
\item \code{neighbors_count}, \code{neighbors_area}, \code{neighbors_percent_shared_edge}, and \code{neighbors_average_aridity}: Number, area sum, proportion of shared edge, and average aridity of the direct neighbors of each ecoregion.
\item \code{climate_aridity_index_average}: Average \href{https://figshare.com/articles/dataset/Global_Aridity_Index_and_Potential_Evapotranspiration_ET0_Climate_Database_v2/7504448}{aridity index} of the ecoregion.
\item \code{climate_bio1_average} to \code{climate_bio15_average}: Bioclimatic variables representing average annual temperature (BIO1), temperature_seasonality (BIO4), average and maximum temperature of the warmest month (BIO5), minimum, average and maximum monthly rainfall (BIO12), and precipitation seasonlality (BIO15).
\item \code{climate_hypervolume}: Volume of the climatic envelope of the ecoregion, computed with the \href{https://cran.r-project.org/package=hypervolume}{hypervolume} package.
\item \code{landcover_bare_percent_average}, \code{landcover_herbs_percent_average}, \code{landcover_trees_percent_average}, and \code{landcover_ndvi_average}: Average cover percentage of bare soil, herbs, and trees, and NDVI, extracted from \href{https://modis-land.gsfc.nasa.gov/vcc.html}{MODIS Vegetation Continuous Fields} and the MODIS NDVI/EVI product.
\item \code{topography_elevation_average} and \code{topography_elevation_range}: Average elevation of the ecoregion and its range.
\item \code{fragmentation_ai} to \code{fragmentation_te}: Geographic fragmentation indices of the ecoregion,  computed with the R package \href{https://CRAN.R-project.org/package=landscapemetrics}{landscapemetrics}.
\item \code{primary_productivity}: Character. Ordered character version of the column \code{landcover_ndvi_average}.
\item \code{landlocked}: Binary. Defines whether a ecoregion is landlocked (value \code{1}) or not (value \code{0}).
\item \code{dominant_landcover}: Factor. Dominant landcover of the ecoregion (one of: "bare", "herbs", and "trees"), computed from the columns \code{landcover_bare_percent_average}, \code{landcover_herbs_percent_average}, and \code{landcover_trees_percent_average}.

}
}
\usage{
data(ecoregions_df)
}
\description{
Richness of vascular plants of the American ecoregions as defined in \href{https://ecoregions2017.appspot.com/}{Ecoregions 2017}, and computed from GBIF data, and environmental predictors.
}
\seealso{
\link{ecoregions_distance_matrix}, \link{ecoregions_continuous_response}, \link{ecoregions_numeric_predictors}
}
\keyword{datasets}
