% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_df.R
\docType{data}
\name{ecoregions_distance_matrix}
\alias{ecoregions_distance_matrix}
\title{Matrix of distances among ecoregion edges.}
\format{
A numeric matrix with 227 rows and columns.
}
\usage{
data(ecoregions_distance_matrix)
}
\description{
Distance matrix (in km) among the edges of the American ecoregions described in the \link{ecoregions_df} dataset.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
