% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_numeric_predictors}
\alias{ecoregions_numeric_predictors}
\title{Names of the numeric predictor variables}
\format{
Character vector with 43 predictor names.
}
\usage{
data(ecoregions_numeric_predictors)
}
\description{
Character vector with the names of the numeric predictors.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
