% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_df.R
\docType{data}
\name{ecoregions_polygons}
\alias{ecoregions_polygons}
\title{Ecoregion polygons.}
\format{
A data frame of the classes "sf" and "data.frame".
}
\usage{
data(ecoregions_polygons)
}
\description{
Simplified ecoregion polygons in sf format.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
