% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions_df.R
\docType{data}
\name{ecoregions_predvar_names}
\alias{ecoregions_predvar_names}
\title{Names of the predictor variables}
\format{
Character vector with 43 predictor names.
}
\usage{
data(ecoregions_predvar_names)
}
\description{
Character vector with the names of the predictors.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
