% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_sf}
\alias{ecoregions_sf}
\title{Ecoregions training data and polygons.}
\format{
A data frame of the classes "sf" and "data.frame".
}
\usage{
data(ecoregions_sf)
}
\description{
Simplified ecoregion polygons in sf format.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
