% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions_tibble}
\alias{ecoregions_tibble}
\title{Ecoregions training tibble.}
\format{
A tibble.
}
\usage{
data(ecoregions_sf)
}
\description{
Simplified ecoregion polygons in sf format.
}
\seealso{
\link{ecoregions_df}
}
\keyword{datasets}
