% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_pca.R
\name{fe_pca_factors}
\alias{fe_pca_factors}
\title{Generates PCA factors to summarize groups of variables}
\usage{
fe_pca_factors(
  data = NULL,
  predictors.names = NULL,
  predictor.groups = NULL,
  replace = FALSE,
  max.explained.variance = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required, data frame, tibble, or sf data frame) Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{predictors.names}{(required; character vector with column names of \code{data}) Character vector with the names of numeric predictors. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictor.groups}{(required, named list of character strings or vectors) List indicating the predictor groups for which PCA factors should be generated. The predictors to use on each group are either indicated by their actual names, or by a pattern. For example, \code{predictor.groups = list(a = c("x", "y", "z"))} generates the PCA factors \code{a_PC1}, \code{a_PC2}, and \code{a_PC3} From the variables \code{"x"}, \code{"y"}, and \code{"z"}. Also, a pattern can be used. For example, \code{predictor.groups = list(climate = "climate")} will generate PCA factors named with the prefix \code{"climate_"} for all the predictors containing the word "climate" in \code{predictors.names}, or the column names of \code{data} if \code{predictors.names = NULL}. If \code{NULL}, then all predictors in \code{predictors.names} are used as group, all columns in \code{data}. Default: \code{NULL}}

\item{replace}{(optional, logical) If \code{TRUE}, the original groups of predictors are replaced in the output data frame with the PCA factors. Default: \code{FALSE}}

\item{max.explained.variance}{(optional, numeric) Maximum cumulative sum of the explained variance of the selected PCA factors. Default: \code{0.80}}

\item{verbose}{(optional, logical) If \code{TRUE}, the function prints message indicating what columns have been scaled. Default: \code{TRUE}}
}
\value{
The input data frame with PCA factors.
}
\description{
The function selects what predictors to use as input for the PCA as follows:
If \code{predictor.groups} is provided, then the predictors listed each slot of the list are used to generate separate sets of PCA factors, and the respective names of this list are used to identify the output PCA factors. If \code{predictor.groups = NULL}, then all predictors in \code{predictors.names} are used as input for the PCA analysis.
}
\examples{

if(interactive()){

data(
  ecoregions_df,
  ecoregions_all_predictors,
  ecoregions_continuous_response
  )

df <- fe_pca_factors(
  data = ecoregions_df,
  predictors.names = ecoregions_all_predictors,
  predictor.groups = list(
    climate = "climate",
    landcover = "landcover",
    fragmentation = "fragmentation"
  ),
  replace = TRUE,
  max.explained.variance = 0.80,
  verbose = TRUE
)

colnames(df)

}
}
