% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_rescale.R
\name{fe_rescale}
\alias{fe_rescale}
\title{Scaling to a new range}
\usage{
fe_rescale(
  data = NULL,
  response.name = NULL,
  predictors.names = NULL,
  new.min = 0,
  new.max = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required, data frame, tibble, or sf data frame) Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{response.name}{(optional; character string) Character string with the name of the response variable. Must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictors.names}{(optional; character vector with column names of \code{data}) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{new.min}{(optional, numeric) New minimum value. Default: \code{0}}

\item{new.max}{(optional, numeric) New maximum value. Default: \code{1}}

\item{verbose}{(optional, logical) If \code{TRUE}, the function prints message indicating what columns have been scaled. Default: \code{TRUE}}
}
\value{
Input data frame with scaled numeric response and predictors.
}
\description{
Re-scales all indicated variables to a new range.
}
\examples{
if(interactive()){

data(
  ecoregions_df,
  ecoregions_all_predictors,
  ecoregions_continuous_response
)

data.rescaled <- fe_rescale(
  data = ecoregions_df,
  response.name = ecoregions_continuous_response,
  predictors.names = ecoregions_all_predictors,
  new.min = 0,
  new.max = 100
  )

}

}
