% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_scale.R
\name{fe_scale}
\alias{fe_scale}
\title{Scaling and centering}
\usage{
fe_scale(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  scale = TRUE,
  center = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required, data frame, tibble, or sf data frame) Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{dependent.variable.name}{(optional; character string) Character string with the name of the response variable. Must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictor.variable.names}{(optional; character vector with column names of \code{data}) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{scale}{(optional, logical) If \code{TRUE}, the data is scaled. If \code{center = TRUE}, the scaling is performed using the expression \code{(a - mean(a))/sd(a)} If \code{center = FALSE}, then a numeric column \code{a} is scaled using the expression \code{a/(sqrt(sum(a^2)/(length(a)-1)))}}

\item{center}{(optional, logical) If \code{TRUE}, the selected numeric columns are centered by subtracting their respective means. Default: \code{TRUE}}

\item{verbose}{(optional, logical) If \code{TRUE}, the function prints message indicating what columns have been scaled. Default: \code{TRUE}}
}
\value{
Input data frame with scaled numeric response and predictors.
}
\description{
Convenient wrapper of the function \code{\link[=scale]{scale()}} to standardize a given set of numeric columns in a training data frame via centering and/or scaling.

Centering:

If \code{center = TRUE} (default value), each element in a numeric column \code{a} is centered by computing \code{a - mean(a)}. This operation centers the values of \code{a} to zero, generating negative values.

Scaling:

If \code{center = TRUE}, the data in column \code{a} is centered via \code{a - mean(a)} and then is divided by the standard deviation of \code{a}, as in \code{(a - mean(a))/sd(a)}.

If \code{center = FALSE}, the column \code{a} is divided by its root-mean-square using the expression \code{a/(sqrt(sum(a^2)/(length(a)-1)))}.

What columns are standardized?

If only \code{data} is provided, then all numeric columns in \code{data} are scaled.

If \code{data} and \code{dependent.variable.name} are provided, then \code{dependent.variable.name} is scaled if it is numeric and continuous (binary dependent variables with values 0 and 1 are not scaled). This is useful to compare the importance scores of models fitted with response variables with different ranges and/or units.

If \code{data} and \code{predictor.variable.names} are provided, then all numeric variables in \code{predictor.variable.names} are scaled.
}
\examples{
if(interactive()){

data(
  ecoregions_df,
  ecoregions_all_predictors,
  ecoregions_continuous_response
)

data.scaled <- fe_scale(
  data = ecoregions_df,
  dependent.variable.name = ecoregions_continuous_response,
  predictor.variable.names = ecoregions_all_predictors
  )

}

}
