% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_scale_predictors.R
\name{fe_scale_all}
\alias{fe_scale_all}
\title{Scale model predictors}
\usage{
fe_scale_all(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL
)
}
\arguments{
\item{data}{(required if \code{model} is \code{NULL}; data frame or tibble) Data frame with a response variable and a set of predictors. If \code{data} is a tibble, all data frames in the output model are coerced to tibble. Default: \code{NULL}}

\item{dependent.variable.name}{(required if \code{model} is \code{NULL}; character string) Character string with the name of the response variable. Must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictor.variable.names}{(required if \code{model} is \code{NULL}; character vector with column names of \code{data}) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Optionally, the result of \code{\link[=mc_auto_cor]{mc_auto_cor()}} or \code{\link[=mc_auto_vif]{mc_auto_vif()}}. Default: \code{NULL}}
}
\value{
Input data frame with scaled numeric response and predictors.
}
\description{
Applies the \code{scale()} function to the numeric predictors of a training data frame.
}
\examples{
if(interactive()){

data(
ecoregions_df,
ecoregions_all_predictors,
ecoregions_continuous_response
)

data.scaled <- fe_scale_all(
  data = ecoregions_df,
  dependent.variable.name = ecoregions_continuous_response,
  predictor.variable.names = ecoregions_all_predictors
  )

}

}
