% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_scale_predictors.R
\name{fe_scale_predictors}
\alias{fe_scale_predictors}
\title{Scale model predictors}
\usage{
fe_scale_predictors(data, predictor.variable.names)
}
\arguments{
\item{data}{(required if \code{model} is \code{NULL}; data frame or tibble) Data frame with a response variable and a set of predictors. If \code{data} is a tibble, all data frames in the output model are coerced to tibble. Default: \code{NULL}}

\item{predictor.variable.names}{(required if \code{model} is \code{NULL}; character vector with column names of \code{data}) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Optionally, the result of \code{\link[=auto_cor]{auto_cor()}} or \code{\link[=auto_vif]{auto_vif()}}. Default: \code{NULL}}
}
\value{
Input data frame with scaled numeric predictors.
}
\description{
Applies the \code{scale()} function to the numeric predictors of a training data frame.
}
\examples{
if(interactive()){

data(
ecoregions_df,
ecoregions_all_predictors
)

data.scaled <- fe_scale_predictors(
  data = ecoregions_df,
  predictor.variable.names = ecoregions_all_predictors
  )

}

}
