% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_spatial_predictors.R
\name{fe_spatial_predictors}
\alias{fe_spatial_predictors}
\title{Adds spatial predictors to a training data frame}
\usage{
fe_spatial_predictors(
  data,
  response.name,
  predictors.names,
  distance.matrix,
  distance.thresholds,
  what.mem = "positive",
  method = "mem",
  cluster = NULL
)
}
\arguments{
\item{data}{(required; data.frame or tibble) A data frame, tibble, or sf. Default: \code{NULL}.}

\item{response.name}{(optional; character string) Name of the dependent variable. Required when there are categorical variables in \code{predictors.names}. Default: \code{NULL}}

\item{predictors.names}{(required; character vector with column names of \code{data}) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{distance.matrix}{(required; distance matrix) Squared matrix with the distances among the records in \code{data}. The number of rows of \code{distance.matrix} and \code{data} must be the same. Default: \code{NULL}}

\item{distance.thresholds}{(optional; numeric vector with distances in the same units as \code{distance.matrix}) Numeric vector with neighborhood distances to be used to compute Moran's Eigenvector Maps. Default: \code{NULL}}

\item{what.mem}{(optional, character) Only relevant for \verb{method = "mem}. Character string indicating what Moran's Eigenvector Maps to generate. If "positive", only MEMs with positive eigenvectors are returned. If "negative", only MEMs with negative eigenvectors are returned. If "all", all MEMs are returned. Default: \code{"positive"}}

\item{method}{(optional, character) Name of the method used to generate spatial predictors. If "hengl", the spatial predictors are the columns of the distance matrix between cases. If "mem", then Moran's Eigenvector Maps are used as spatial predictors.}

\item{cluster}{(optional, only relevant for \code{method = "mem"}; cluster object) A cluster definition generated with \code{parallel::makeCluster()} or \code{\link{start_cluster}}. Only advisable if you need to spread a large number of repetitions over the nodes of a large cluster when working with large data. The function does not stop a cluster, please remember to shut it down with \code{parallel::stopCluster(cl = cluster_name)} or \code{spatialRF::stop_cluster()} at the end of your pipeline. Default: \code{NULL}}
}
\value{
Adds new columns to the \code{data} argument. If \code{method = "hengl"}, then these columns are named "distance_from_X", where X is the index of the case the distance is being measured from. If \code{method = "mem"}, then these columns are named "mem_D_X", where D is the distance threshold for which the MEM was computed, and X is the index is the index of the case the MEM has been computed for.
}
\description{
Adds spatial predictors to a training data frame.

Spatial predictors represent the spatial structure of the training data, and are only required when model residuals show spatial autocorrelation. Spatial predictors are a proxy for spatial processes that might not be represented in other predictors.

When using the method "hengl", the distance matrix is added as spatial component to the trainng data frame.

When using the method "mem", then Moran's Eigenvector Maps (MEMs) are added as spatial component instead. Using the argument \code{what.mem} the user can select whether to add only positive, only negative, or all MEMs to the input data frame.
}
\examples{
if(interactive()){{

data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_all_predictors,
  ecoregions_continuous_response
 )

 #speed up MEM selection
 cluster <- start_cluster()

 ecoregions_df <- fe_spatial_predictors(
   data = ecoregions_df,
   response.name = ecoregions_continuous_response,
   predictors.names,
   distance.matrix = ecoregions_distance_matrix,
   what.mem = "positive",
   method = "mem",
   cluster = cluster
   )

   stop_cluster(cluster)

}}
}
