% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evaluation.R
\name{get_evaluation}
\alias{get_evaluation}
\title{get_evaluation}
\usage{
get_evaluation(x)
}
\arguments{
\item{x}{A model resulting from \code{\link[=rf_evaluate]{rf_evaluate()}}}
}
\value{
A data frame.
}
\description{
Returns an evaluation performed with \code{\link[=rf_evaluate]{rf_evaluate()}} as a data frame.
}
