% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interactions.R
\name{get_interactions}
\alias{get_interactions}
\title{Extract interaction details}
\usage{
get_interactions(model, type = c("selected", "screening", "all"))
}
\arguments{
\item{model}{Model from \code{\link[=rf_interactions]{rf_interactions()}}}

\item{type}{Character: "selected", "screening", or "all". Default: "selected"}
}
\value{
Data frame or list with interaction details
}
\description{
Extract interaction selection details from rf_interactions model
}
\examples{
if (interactive()) {
  data(plants_df, plants_xy, plants_rf)

  # Fit model with interactions
  m_int <- rf_interactions(
    model = plants_rf,
    xy = plants_xy,
    repetitions = 5,
    n.cores = 1
  )

  # Get selected interactions
  get_interactions(m_int, type = "selected")

  # Get all tested interactions
  get_interactions(m_int, type = "screening")

  # Get both
  get_interactions(m_int, type = "all")
}
}
\concept{get_functions}
