% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster.R
\name{make_cluster}
\alias{make_cluster}
\title{Defines a cluster}
\usage{
make_cluster(
  cluster.ips = NULL,
  cluster.cores = parallel::detectCores() - 1,
  cluster.user = Sys.info()[["user"]],
  cluster.port = "11000",
  outfile = NULL
)
}
\arguments{
\item{cluster.ips}{Character vector with the IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed. Default: \code{NULL}.}

\item{cluster.cores}{Numeric integer vector, number of cores on each machine. Default: \code{parallel::detectCores() - 1}.}

\item{cluster.user}{Character string, name of the user (should be the same throughout machines), Defaults to the current system user.}

\item{cluster.port}{Character, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: \code{"11000"}}

\item{outfile}{Where to direct the messages provided by the workers. When working on a local computer, \code{""} prints the worker's messages in the console. A text file path will append worker's messages on the given file.  Default: \verb{/dev/null} en Linux and  \verb{nul:} on windows.}
}
\value{
A list ready to be used as input for the \code{spec} argument of the function \link[parallel]{makeCluster}.
}
\description{
Defines a cluster. With default arguments it returns a cluster defined in the local machine. If the arguments \code{cluster.ips} and \code{cluster.cores} are provided, then a Beowulf cluster spanning as many machines as IPs is returned. All functions taking clusters as inputs in this package will register the cluster themselves using \code{doParallel::registerDoParallel()} internally, but if you are going to use this function for other purposes, please register your cluster before using it. Notice that Beowulf clusters require a particular configuration for your local network (see section \verb{Network settings} at \href{https://www.blasbenito.com/post/01_home_cluster/}{https://www.blasbenito.com/post/01_home_cluster/}).
}
\examples{
if(interactive()){

local.cluster <- make_cluster(cluster.cores = 1)
#this cluster is registered by the packaage functions when used
stop_cluster(cluster = local.cluster)

beowulf.cluster <- make_cluster(
  cluster.ips = c(
   "10.42.0.1",
   "10.42.0.34",
   "10.42.0.104"
   ),
 cluster.cores = c(1, 1, 1),
 cluster.user = "blas",
 cluster.port = "11000"
)

stop_cluster(cluster = beowulf.cluster)
}

}
