% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_auto_cor.R
\name{mc_auto_cor}
\alias{mc_auto_cor}
\title{Automated multicollinearity reduction via pairwise Pearson correlation}
\usage{
mc_auto_cor(
  data = NULL,
  predictors.names = NULL,
  response.name = NULL,
  preference.order = NULL,
  max.cor = 0.75,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{predictors.names}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data except \code{response.name} are used. Default: \code{NULL}}

\item{response.name}{(optional; character string) Name of the dependent variable. Only required when there are categorical variables within \code{predictors.names}. Default: \code{NULL}}

\item{preference.order}{(optional; character vector) Character vector indicating the user's order of preference to keep variables. Predictors not included in this argument are ranked by the sum of their correlation with other variables (variables with higher sums receive lower ranks and have therefore lower preference order). Default: \code{NULL}.}

\item{max.cor}{(optional; numeric) Numeric between 0 and 1. Maximum Pearson correlation between any pair of the selected variables. Higher values return larger number of predictors with higher multicollinearity. Default: \code{0.75}}

\item{verbose}{(optional, logical) Logical. if \code{TRUE}, describes the function operations to the user. Default: \code{TRUE}}
}
\value{
Character vector with the names of uncorrelated predictors.
}
\description{
Iterative reduction of multicollinearity via Pearson bivariate correlation.

The function \code{mc_auto_cor()} applies a recursive algorithm to remove variables with a Pearson correlation with another variable higher than a given threshold (defined by the argument \code{max.cor}).  When two variables are correlated above this threshold, the one with the highest sum of R-squared with all the other variables is removed.

The function \code{mc_auto_cor()} allows the user to define preference selection order via the argument \code{preference.order}. If \code{preference.order} is \verb{"c("a", "b")}, they are correlated above \code{max.cor}, there there are other variables in \code{data}, and the sum of R-squared with all other variables of \code{"a"} is higher than the sum of \code{"b"}, then \code{"b"} is removed anyway.

The argument \code{preference.order} allows the user to "protect" variables that might be interesting or even required for the given analysis.

If \code{preference.order} is not provided, then the predictors are ranked from lower to higher sum of R-squared with the other preodictors, and removed one by one until the maximum R-squared of the correlation matrix is lower than \code{max.cor}.

If there are categorical variables named in \code{predictors.names} and \code{response.name} is provided, then the function applies \code{\link[=fe_target_encoding]{fe_target_encoding()}} with the method "mean" to transform the categorical variables into numeric. If a categorical variable is selected, then its original categorical values are returned.

Please note that near-zero variance columns are ignored by this function.
}
\examples{
if(interactive()){

 #loading example data
 data(
  ecoregions_df,
  ecoregions_numeric_predictors
  )

 #on a data frame
 out <- mc_auto_cor(
   data = ecoregions_df,
   predictors.names = ecoregions_numeric_predictors
 )

 #getting the correlation matrix
 out$cor

 #getting the names of the selected variables
 out$selected.variables

 #getting the data frame of selected variables
 out$selected.variables.df

 #with preference order (fiver first in ecoregions_numeric_predictors)
 out <- mc_auto_cor(
   data = ecoregions_df,
   predictors.names = ecoregions_numeric_predictors,
   preference.order = ecoregions_numeric_predictors[1:5],
 )

 #with pipes
 out <- mc_auto_cor(
   data = ecoregions_df,
   predictors.names = ecoregions_numeric_predictors
 ) \%>\%
 mc_auto_vif()

}
}
\seealso{
\code{\link[=mc_auto_vif]{mc_auto_vif()}}
}
