% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_multithreshold.R
\name{moran_multithreshold}
\alias{moran_multithreshold}
\title{moran_multithreshold}
\usage{
moran_multithreshold(
  x = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{numeric vector, generally a model residuals, Default: NULL}

\item{distance.matrix}{distance matrix among the records represented in the numeric vector. The number of rows of this matrix must be equal to the length of x. Default: NULL}

\item{distance.thresholds}{numeric vector, distances below each value in the distance matrix are set to 0 for the computation of Moran's I. If NULL, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: NULL}

\item{plot}{boolean, plots Moran's I values for each distance threshold if TRUE, Default: TRUE}
}
\value{
a named list with the slots:
\describe{
\item{df}{dataframe with the results of \link{moran} per distance threshold.}
\item{plot}{a plot of Moran's I across distance thresholds.}
\item{max.moran}{maximum value of Moran's I across thresholds.}
\item{max.moran.distance.threshold}{distance threshold with the maximum Moran's I value.}
}
}
\description{
Applies \link{moran} to different distance thresholds at the same time, and plots Moran's I value for each distance threshold.
}
\details{
The idea behind this function comes from the uncertainty about what "neighborhood" means in ecological systems (1000km in geological time mean little, but 100m might be quite a long distance for a tree to disperse seeds over), and allows to explore spatial autocorrelation of model residuals for several minimum-distance criteria at once.
}
\examples{
\dontrun{
if(interactive()){
 data(distance_matrix)
 data(plant_richness)
 moran.out <- moran_multithreshold(
   x = plant_richness$richness_species_vascular,
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 100, 1000, 10000),
   plot = TRUE
   )
 moran.out
 }
}
}
