% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_size.R
\name{objects_size}
\alias{objects_size}
\title{objects_size}
\usage{
objects_size(n = 10)
}
\arguments{
\item{n}{number of objects to show, Default: 10}
}
\value{
a data frame with the row names indicating the object name, the field 'Type' indicating the object type, 'Size' indicating the object size, and the columns 'Length/Rows' and 'Columns' indicating the object dimensions.
}
\description{
shows the size of the objects currently in the R environment. Helper function to locate large objects cluttering the R environment.
}
\examples{
 x <- matrix(runif(100), 10, 10)
 y <- matrix(runif(10000), 100, 100)
 objects_size()
}
