% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_function.R
\name{optimization_function}
\alias{optimization_function}
\title{optimization_function}
\usage{
optimization_function(
  x = NULL,
  weight.r.squared = NULL,
  weight.penalization.n.predictors = NULL,
  optimization.method = "moran.i"
)
}
\arguments{
\item{x}{Optimization data frame generated internally by \link{select_spatial_predictors_optimized} and \link{select_spatial_predictors_optimized}. Default: NULL}

\item{weight.r.squared}{numeric between 0 and 1, weight of R-squared in the optimization process. Default: NULL}

\item{weight.penalization.n.predictors}{numeric between 0 and 1, weight of the penalization on the number of added spatial predictors. Default: NULL}

\item{optimization.method}{character, one of "moran.i", and "p.value". Default: 'moran.i'}
}
\value{
A numeric vector with the optimization criteria.
}
\description{
optimizes the selection of spatial predictors using two different methods: "moran.i", and "p.value". The user shouldn't use this function. See Details.
}
\details{
The method "moran.i" tries to maximize \verb{1 - Moran's} I while taking into account the R-squared of the model and a penalization on the number of introduced spatial predictors through the expression \verb{(1 - Moran's I) + w1 * r.squared - w2 * penalization}. The method "p.value" uses a binary version of the p-values of Moran's I (1 if >= 0.05, 0 otherwise), and uses the expression \verb{max(1 - Moran's I, binary p-value) + w1 * r.squared - w2 * penalization}. The "moran.i" method allows to include more spatial predictors than "p.value"
}
