% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_richness.R
\docType{data}
\name{plant_richness_sf}
\alias{plant_richness_sf}
\title{Centroids of North and South American ecoregions.

Centroids of the American ecoregions represented in \link{plant_richness_df} in format \link[sf]{sf}, with latitude and longitude coordinates and datum WGS84 stored in the column \emph{geom_centroids}.
plant_richness_sf}
\format{
A spatial data frame (sf) with 227 rows
}
\usage{
data(plant_richness_sf)
}
\description{
Centroids of North and South American ecoregions.

Centroids of the American ecoregions represented in \link{plant_richness_df} in format \link[sf]{sf}, with latitude and longitude coordinates and datum WGS84 stored in the column \emph{geom_centroids}.
plant_richness_sf
}
\seealso{
\link{plant_richness_df}, \link{distance_matrix}
}
\keyword{datasets}
