% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_distance}
\alias{plants_distance}
\title{Matrix of distances among ecoregion edges.}
\format{
A numeric matrix with 227 rows and columns.
}
\usage{
data(plants_distance)
}
\description{
Distance matrix (in km) among the edges of the American ecoregions described in the \link{plants_df} dataset.
}
\keyword{datasets}
