% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_predictors}
\alias{plants_predictors}
\title{Predictor variable names for plant richness examples}
\format{
A character vector of length 17.
}
\usage{
data(plants_predictors)
}
\description{
Character vector containing the names of predictor variables in \link{plants_df} (columns 5 to 21).
}
\keyword{datasets}
