% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_response}
\alias{plants_response}
\title{Response variable name for plant richness examples}
\format{
A character string of length 1.
}
\usage{
data(plants_response)
}
\description{
Character string containing the name of the response variable in \link{plants_df}: "richness_species_vascular".
}
\keyword{datasets}
