% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_xy}
\alias{plants_xy}
\title{Coordinates of plant richness data}
\format{
A data frame with 227 rows and 2 columns:
\itemize{
\item \code{x}: Longitude in degrees (WGS84).
\item \code{y}: Latitude in degrees (WGS84).
}
}
\usage{
data(plants_xy)
}
\description{
Data frame containing the x and y coordinates extracted from \link{plants_df} for use in spatial modeling functions.
}
\keyword{datasets}
