% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{plot_interactions}
\alias{plot_interactions}
\title{Plot variable interactions}
\usage{
plot_interactions(model, which = "all", verbose = TRUE)
}
\arguments{
\item{model}{Model from \code{\link[=rf_interactions]{rf_interactions()}}}

\item{which}{Character vector of interaction names, or "all" (default)}

\item{verbose}{Logical. If \code{TRUE}, the plot is printed. Default: \code{TRUE}}
}
\value{
ggplot or patchwork object
}
\description{
Plot selected interactions from rf_interactions model
}
\examples{
if (interactive()) {
  data(plants_df, plants_xy, plants_rf)

  # Fit model with interactions
  m_int <- rf_interactions(
    model = plants_rf,
    xy = plants_xy,
    repetitions = 5,
    n.cores = 1
  )

  # Plot all interactions
  plot_interactions(m_int)

  # Plot specific interactions
  int_names <- m_int$interactions$selected$interaction.name[1:2]
  plot_interactions(m_int, which = int_names)
}
}
\concept{plot_functions}
