% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moran.R
\name{plot_moran}
\alias{plot_moran}
\title{plot_moran}
\usage{
plot_moran(x, verbose = TRUE)
}
\arguments{
\item{x}{a data frame resulting from \link{moran_multithreshold}. Default: NULL}

\item{verbose}{logical, if TRUE, the resulting table is printed into the console, Default: TRUE}
}
\value{
a ggplot.
}
\description{
Plots results of spatial autocorrelation tests for a variety of functions within the package.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)
 data(distance.matrix)
 #basic model
 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000),
   verbose = FALSE
 )

 plot_moran(x = rf.model$spatial.correlation.residuals$per.distance)
 }
}
}
