% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization}
\alias{plot_optimization}
\title{Optimization plot of a selection of spatial predictors}
\usage{
plot_optimization(
  model,
  point.color = viridis::viridis(
    100,
    option = "F",
    direction = -1
  ),
  verbose = TRUE
)
}
\arguments{
\item{model}{A model produced by \code{\link[=rf_spatial]{rf_spatial()}}, or an optimization data frame produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}} or \code{\link[=select_spatial_predictors_recursive]{select_spatial_predictors_recursive()}}.}

\item{point.color}{Colors of the plotted points. Can be a single color name (e.g. "red4"), a character vector with hexadecimal codes (e.g. "#440154FF" "#21908CFF" "#FDE725FF"), or function generating a palette (e.g. \code{viridis::viridis(100)}). Default: \code{viridis::viridis(100, option = "F", direction = -1)}}

\item{verbose}{Logical, if \code{TRUE} the plot is printed. Default: \code{TRUE}}
}
\value{
A ggplot.
}
\description{
Plots optimization data frames produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}}
and \code{\link[=select_spatial_predictors_recursive]{select_spatial_predictors_recursive()}}.
}
\details{
If the method used to fit a model with \code{\link[=rf_spatial]{rf_spatial()}} is "hengl", the function returns nothing, as this method does not require optimization.
}
\examples{
if(interactive()){

 #loading example data
#loading example data
data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_numeric_predictors,
  ecoregions_continuous_response
  )

 #spatial model
 rf.model <- rf_spatial(
   data = ecoregions_df,
   response.name = ecoregions_continuous_response,
   predictors.names = ecoregions_numeric_predictors,
   distance.matrix = ecoregions_distance_matrix,
   distance.thresholds = 0,
   method = "mem.moran.sequential",
   n.cores = 1,
   seed = 1
 )

 #plotting selection of spatial predictors
 plot_optimization(model = rf.model)


}
}
