% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals_diagnostics.R
\name{plot_residuals_diagnostics}
\alias{plot_residuals_diagnostics}
\title{Plot residuals diagnostics}
\usage{
plot_residuals_diagnostics(model)
}
\arguments{
\item{model}{A model produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}.}
}
\value{
A patchwork object.
}
\description{
Plots normality and autocorrelation tests (if available) of a model's residuals.
}
\examples{
\dontrun{
if(interactive()){

 data(plant_richness_df)
 data(distance_matrix)

 x <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix
 )

 plot_residuals_diagnostics(x)

 }
}
}
