% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_importance_spatial.R
\name{prepare_importance_spatial}
\alias{prepare_importance_spatial}
\title{Prepares variable importance objects for spatial models}
\usage{
prepare_importance_spatial(model)
}
\arguments{
\item{model}{An importance data frame with spatial predictors, or a model fitted with \code{\link[=rf_spatial]{rf_spatial()}}.}
}
\value{
A list with importance data frames in different formats depending on whether the model was fitted with \code{\link[=rf]{rf()}} or \code{\link[=rf_repeat]{rf_repeat()}}.
}
\description{
Prepares variable importance data frames and plots for models fitted with \code{\link[=rf_spatial]{rf_spatial()}}.
}
\examples{
if(interactive()){

#loading example data
data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_predvar_names,
  ecoregions_depvar_name
  )

 #fittind spatial model
 model <- rf_spatial(
   data = ecoregions_df,
   dependent.variable.name = ecoregions_depvar_name,
   predictor.variable.names = ecoregions_predvar_names,
   distance.matrix = ecoregions_distance_matrix,
   distance.thresholds =  0,
   n.cores = 1
 )

 #preparing the importance data frame
 importance <- prepare_importance_spatial(model)
 names(importance)

}
}
