% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_evaluation.R, R/print_functions.R
\name{print_evaluation}
\alias{print_evaluation}
\title{Prints cross-validation results}
\usage{
print_evaluation(model)

print_evaluation(model)
}
\arguments{
\item{model}{A model resulting from \code{\link[=rf_evaluate]{rf_evaluate()}}.}
}
\value{
A table printed to the standard output.

A table printed to the standard output.
}
\description{
Prints the results of an spatial cross-validation performed with \code{\link[=rf_evaluate]{rf_evaluate()}}.

Prints the results of an spatial cross-validation performed with \code{\link[=rf_evaluate]{rf_evaluate()}}.
}
\examples{
if(interactive()){

#loading example data
data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_predictor_variable_names,
  ecoregions_dependent_variable_name
  )

 #fitting random forest model
 rf.model <- rf(
   data = ecoregions_df,
   dependent.variable.name = ecoregions_dependent_variable_name,
   predictor.variable.names = ecoregions_predictor_variable_names,
   distance.matrix = ecoregions_distance_matrix,
   distance.thresholds = 0,
   n.cores = 1,
   verbose = FALSE
 )

#evaluation with spatial cross-validation
rf.model <- rf_evaluate(
  model = rf.model,
  xy = plant_richness_df[, c("x", "y")],
  n.cores = 1
)

#checking evaluation results
print_evaluation(rf.model)

}
if(interactive()){

#loading example data
data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_predictor_variable_names,
  ecoregions_dependent_variable_name
  )

 #fitting random forest model
 rf.model <- rf(
   data = ecoregions_df,
   dependent.variable.name = ecoregions_dependent_variable_name,
   predictor.variable.names = ecoregions_predictor_variable_names,
   distance.matrix = ecoregions_distance_matrix,
   distance.thresholds = 0,
   n.cores = 1,
   verbose = FALSE
 )

#evaluation with spatial cross-validation
rf.model <- rf_evaluate(
  model = rf.model,
  xy = plant_richness_df[, c("x", "y")],
  n.cores = 1
)

#checking evaluation results
print_evaluation(rf.model)

}
}
\seealso{
\code{\link[=plot_evaluation]{plot_evaluation()}}

\code{\link[=plot_evaluation]{plot_evaluation()}}
}
