% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curves}
\alias{response_curves}
\title{response_curves}
\usage{
response_curves(
  model = NULL,
  variables = NULL,
  quantiles = c(0.1, 0.5, 0.9),
  grid.resolution = 100,
  show.data = FALSE
)
}
\arguments{
\item{model}{A model fitted with \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}.}

\item{variables}{Character vector, names of predictors to plot. If \code{NULL}, the most important variables (importance higher than the median) in \code{model} are selected. Default: \code{NULL}.}

\item{quantiles}{Numeric vector with values between 0 and 1, argument \emph{probs} of \link[stats]{quantile}. Quantiles to set the other variables to. Default: c(0.1, 0.5, 0.9)}

\item{grid.resolution}{Integer between 20 and 500. Resolution of the plotted curve Default: 100.}

\item{show.data}{Logical, if TRUE, the observed data is plotted along with the response curves. Default. FALSE}
}
\value{
A list with slots named after the selected \code{variables} with one ggplot each.
}
\description{
Plots response surfaces for a given pair of predictors from an \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or an \code{\link[=rf_spatial]{rf_spatial()}} model.
}
\details{
All variables that are not plotted are set to the values of their respective quantiles to plot the response surfaces. The output list can be plotted all at once with \code{patchwork::wrap_plots(p)} or \code{cowplot::plot_grid(plotlist = p)}, or one by one by extracting each plot with print(p[\link{1}]).
}
\examples{
\dontrun{
if(interactive()){
data(plant_richness_df)

m <- rf(
 data = plant_richness_df,
 dependent.variable.name = "richness_species_vascular",
 predictor.variable.names = colnames(plant_richness_df)[5:21],
 verbose = FALSE
)

p <- response_curves(model = m)
 }
}
}
