% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_interactions.R
\name{rf_interactions}
\alias{rf_interactions}
\title{Suggest variable interactions for random forest models}
\usage{
rf_interactions(
  model = NULL,
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  ranger.arguments = NULL,
  trees.per.variable = NULL,
  importance.threshold = NULL,
  repetitions = 5,
  verbose = TRUE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = 11000
)
}
\arguments{
\item{model}{A model fitted with \code{\link[=rf]{rf()}}. If used, the arguments \code{data}, \code{dependent.variable.name}, \code{predictor.variable.names}, \code{distance.matrix}, \code{distance.thresholds}, \code{ranger.arguments}, \code{trees.per.variable}, and \code{scaled.importance} are taken directly from the model definition. Default: \code{NULL}}

\item{data}{Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{dependent.variable.name}{Character string with the name of the response variable. Must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictor.variable.names}{Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{ranger.arguments}{Named list with \link[ranger]{ranger} arguments (other arguments of this function can also go here). All \link[ranger]{ranger} arguments are set to their default values except for 'importance', that is set to 'permutation' rather than 'none'. Please, consult the help file of \link[ranger]{ranger} if you are not familiar with the arguments of this function.}

\item{trees.per.variable}{Number of individual regression trees to fit per variable in \code{predictor.variable.names}. This is an alternative way to define ranger's \code{num.trees}. If \code{NULL}, \code{num.trees} is 500. Default: \code{NULL}}

\item{importance.threshold}{Value of variable importance from \code{model} used as threshold to select variables to generate candidate interactions. Default: Median of the variable importance in \code{model}.}

\item{repetitions}{Integer, number of random forest models to fit in order to assess the importance of the interaction. Default: \code{5}}

\item{verbose}{Logical If \code{TRUE}, messages and plots generated during the execution of the function are displayed, Default: \code{TRUE}}

\item{n.cores}{Integer, number of cores to use during computations. If \code{NULL}, all cores but one are used, unless a cluster is used. Default = \code{NULL}}

\item{cluster.ips}{Character vector with the IPs of the machines in a cluster. The machine with the first IP will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{Numeric integer vector, number of cores to use on each machine.}

\item{cluster.user}{Character string, name of the user (should be the same throughout machines). Defaults to the current system user.}

\item{cluster.port}{Integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: \code{11000}}
}
\value{
A list with two slots: \code{screening}, with the complete screening results; \code{selected}, with the names and the R squared improvement produced by each variable interaction; \code{columns}, data frame with the interactions computed from the data in \code{model$ranger.arguments} after scaling it with \code{\link[=scale_robust]{scale_robust()}}. Variable interactions are computed as \code{a * b} on the scaled data.
}
\description{
Suggests candidate variable interactions by selecting the variables above a given importance threshold (given by the argument \code{importance.threshold}) from a model and combining them in pairs through multiplication (\code{a * b}).

For each variable interaction, a model including all the predictors plus the interaction is fitted, and it's R squared is compared with the R squared of a model fitted without any interactions. This model without interactions can either be provided through the argument \code{model}, or fitted with \code{\link[=rf_repeat]{rf_repeat()}} by providing the arguments \code{data}, \code{dependent.variable.name}, and \code{predictor.variable.names}.

I advise the users not to use variable interactions blindly. Most likely, only one or a few of the suggested interactions may make sense from a domain expertise standpoint.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)

 interactions <- rf_interactions(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   verbose = TRUE
 )

 interactions$screening
 interactions$selected
 interactions$columns

}
}
}
