% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{roc_curve}
\alias{roc_curve}
\title{ROC curve}
\usage{
roc_curve(o, p)
}
\arguments{
\item{o}{Numeric vector with observations, must have the same length as \code{p}.}

\item{p}{Numeric vector with predictions, must have the same length as \code{o}.}
}
\value{
A data frame with the following columns:
\itemize{
\item prediction_threshold: prediction thresholds from 0 to 1.
\item true_positives: proportion of correctly predicted ones per threshold.
\item false_positives: proportion of incorrectly predicted ones per threshold.
\item false_negatives: proportion of correctly predicted zeroes per threshold.
\item true_negatives: proportion of incorrectly predicted zeroes per threshold.
\item sensitivity: probability of predicting a one per threshold.
\item specificity: probability of predicting a zero per threshold.
}
}
\description{
Computes tROC curve in models with binary responses.
}
\examples{
if(interactive()){

 out <- roc_curve(
   o = c(0, 0, 1, 1),
   p = c(0.1, 0.6, 0.4, 0.8)
   )

}
}
