% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_spatial_predictors_optimized.R
\name{select_spatial_predictors_optimized}
\alias{select_spatial_predictors_optimized}
\title{select_spatial_predictors_optimized}
\usage{
select_spatial_predictors_optimized(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  ranger.arguments = NULL,
  spatial.predictors.df = NULL,
  spatial.predictors.ranking = NULL,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = NULL,
  cluster.port = 11000
)
}
\arguments{
\item{data}{(required) data frame with a response variable and a set of (preferably uncorrelated) predictors, Default: NULL}

\item{dependent.variable.name}{(required) string with the name of the response variable. Must be in the column names of 'data', Default: NULL}

\item{predictor.variable.names}{(required) character vector with the names of the predictive variables. Every element must be in the column names of 'data', Default: NULL}

\item{distance.matrix}{(optional) a squared matrix with the distances among the records in 'data'. Notice that the rows of 'distance.matrix' and 'data' must be the same. If not provided, the computation of the Moran's I of the residuals is ommited. Default: NULL.}

\item{distance.thresholds}{(optional) numeric vector, distances below each value in the distance matrix are set to 0 for the computation of Moran's I. If NULL, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: NULL.}

\item{ranger.arguments}{list with \link[ranger]{ranger} arguments. See \link{rf} or \link{rf_repeat} for further details.}

\item{spatial.predictors.df}{data frame of spatial predictors, either a distance matrix, or the PCA factors of the distance matrix produced by \link{pca_multithreshold}.}

\item{spatial.predictors.ranking}{ranking of predictors returned by \link{rank_spatial_predictors}.}

\item{n.cores}{number of cores to use to compute repetitions. If NULL, all cores but one are used, unless a cluster is used.}

\item{cluster.ips}{character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{numeric integer vector, number of cores on each machine.}

\item{cluster.user}{character string, name of the user (should be the same throughout machines), Defaults to the current system user. Default: user name of the current session.}

\item{cluster.port}{integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: 11000.}
}
\value{
a list with two slots: \code{optimization}, a data frame with the index of the spatial predictor added on each iteration, the spatial correlation of the model residuals, and the R-squared of the model, and \code{best.spatial.predictors}, that is a character vector with the names of the spatial predictors that minimize the Moran's I of the residuals and maximize the R-squared of the model.
}
\description{
selects spatial predictors (either a distance matrix or the PCA factors of a distance matrix) following these steps:
\enumerate{
\item Gets the best spatial predictor yielded by \link{rank_spatial_predictors} and fits a model of the form \code{y ~ predictors + best_spatial_predictor_1}. The Moran's I of the residuals of this model are used as reference value for the next step.
\item The remaining spatial predictors are introduced again into \link{rank_spatial_predictors}, and the best one is introduced in a new model \code{y ~  predictors + best_spatial_predictor_1 + best_spatial_predictor_2}.
\item Steps 1 and 2 are repeated until there are no more spatial predictors left.
}
This method allows to select the set of spatial predictors that have the largest joint effect in reducing the spatial correlation of the model residuals while maintaining an R-squared as high as possible. As a consequence of running \link{rank_spatial_predictors} on each iteration, this method includes in the final model less spatial predictors than \link{select_spatial_predictors_sequential} while minimizing spatial correlation and maximizing R-squared as much as possible.
}
\details{
How does the algorithm work? For example, if the function \link{rank_spatial_predictors} returns 10 ranked spatial predictors (sp1 to sp10, being sp7 the best one), \link{select_spatial_predictors_optimized} is going to first fit the model \code{y ~ predictors + sp7}. Then, the spatial predictors sp2 to sp9 are again ranked with \link{rank_spatial_predictors} using the model \code{y ~ predictors + sp7} as reference (at this stage, some of the spatial predictors might be dropped due to lack of effect). When the new ranking of spatial predictors is ready (let's say they are sp5, sp3, and sp4), the best one (sp5) is included in the model \code{y ~ predictors + sp7 + sp5}, and the remaining ones go again to \link{rank_spatial_predictors} to repeat the process until spatial predictors are depleted.
}
\examples{
\dontrun{
if(interactive()){
data("distance_matrix")
data("plant_richness_df")

#common arguments
data = plant_richness_df
dependent.variable.name = "richness_species_vascular"
predictor.variable.names = colnames(plant_richness_df)[5:21]
distance.matrix = distance_matrix
distance.thresholds = c(0, 100, 1000)

#non-spatial model
model <- rf(
  data = data,
  dependent.variable.name = dependent.variable.name,
  predictor.variable.names = predictor.variable.names,
  distance.matrix = distance.matrix,
  distance.thresholds = distance.thresholds
)

#preparing spatial predictors
spatial.predictors <- pca_distance_matrix(
  x = distance.matrix,
  distance.thresholds = distance.thresholds
)

#ranking spatial predictors
spatial.predictors.ranking <- rank_spatial_predictors(
  data = data,
  dependent.variable.name = dependent.variable.name,
  predictor.variable.names = predictor.variable.names,
  spatial.predictors.df = spatial.predictors,
  ranking.method = "moran.i.reduction",
  reference.moran.i = model$spatial.correlation.residuals$max.moran,
  distance.matrix = distance.matrix,
  distance.thresholds = distance.thresholds,
  n.cores = 1,
  multicollinearity.filter = "vif"
)

#selecting the best subset of predictors
selection <- select_spatial_predictors_optimized(
  data = data,
  dependent.variable.name = dependent.variable.name,
  predictor.variable.names = predictor.variable.names,
  distance.matrix = distance.matrix,
  distance.thresholds = distance.thresholds,
  spatial.predictors.df = spatial.predictors,
  spatial.predictors.ranking = spatial.predictors.ranking,
  n.cores = 1
)

selection$optimization
selection$best.spatial.predictors
 }
}
}
