% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_mode.R
\name{statistical_mode}
\alias{statistical_mode}
\title{statistical_mode}
\usage{
statistical_mode(x)
}
\arguments{
\item{x}{numeric or character vector}
}
\value{
statistical mode of x
}
\description{
computes the mode of a numeric or character vector
}
\examples{
\dontrun{
if(interactive()){
 out <- statistical_mode(c(rep(10, 10), rep(9, 9)))
 out
 }
}
}
