% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_cluster.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stops clusters}
\usage{
stop_cluster(cluster = NULL)
}
\arguments{
\item{cluster}{a cluster created with \code{parallel::makeCluster()} or \code{spatialRF::make_cluster()}. Default: \code{NULL}}
}
\value{
This function returns nothing.
}
\description{
Stops and removes any object with the class "cluster" in the global environment by default, or a given named cluster when the argument \code{cluster} is provided.
}
\examples{

#removing one given cluster
cluster <- spatialRF::make_cluster(cluster.cores = 1)
spatialRF::stop_cluster(cluster = cluster)

#removing several clusters
cl1 <- spatialRF::make_cluster(cluster.cores = 1)
cl2 <- spatialRF::make_cluster(cluster.cores = 1)
spatialRF::stop_cluster()
}
