% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_interactions.R
\name{suggest_interactions}
\alias{suggest_interactions}
\title{suggest_interactions}
\usage{
suggest_interactions(
  model = NULL,
  importance.threshold = NULL,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = 11000
)
}
\arguments{
\item{model}{(optional) a model produced by \link{rf}. If used, the arguments \code{data}, \code{dependent.variable.name}, \code{predictor.variable.names}, \code{distance.matrix}, \code{distance.thresholds}, \code{ranger.arguments}, \code{trees.per.variable}, and \code{scaled.importance} are taken directly from the model definition. Default: \code{NULL}}

\item{importance.threshold}{Value of variable importance from \code{model} used as threshold to select variables to generate candidate interactions. Default: Median of the variable importance in \code{model}.}

\item{n.cores}{number of cores to use to compute repetitions. If NULL, all cores but one are used, unless a cluster is used.}

\item{cluster.ips}{character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{numeric integer vector, number of cores on each machine.}

\item{cluster.user}{character string, name of the user (should be the same throughout machines), Defaults to the current system user. Default: user name of the current session.}

\item{cluster.port}{integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: 11000.}
}
\value{
A list with two slots: \emph{selected}, with the names and the R squared improvement produced by each variable interaction; \emph{df}, data frame with the interactions computed from the data in \code{model$ranger.arguments} after scaling it with \code{\link[=scale_robust]{scale_robust()}}. Variable interactions are computed as \code{x * y} on the scaled data.
}
\description{
Takes a model fitted with \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}} to find variable interactions improving the R squared of the model.
}
\details{
This function suggests variable interactions that could improve a model by adding them one at a time to the original \code{model} and assessing their contribution to the model's R squared. Only interactions with a positive contribution are suggested. We advise the users not to use variable interactions blindly. Most likely, only one or a few of the suggested interactions may make sense from a domain expertise standpoint.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)
data(distance_matrix)

rf.model <- rf(
 data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  distance.matrix = distance_matrix,
  distance.thresholds = c(0, 1000, 2000),
  verbose = FALSE
)

interactions <- suggest_interactions(model = rf.model)
interactions$selected
interactions$df
 }
}
}
