% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{vif}
\usage{
vif(x)
}
\arguments{
\item{x}{data frame with numeric columns, typically with a set of predictors.}
}
\value{
a data frame with two columns having the name of the variables in 'x' and their respective VIF values.
}
\description{
computes the variance inflation factor (VIF) of the colums in a data frame.
}
\examples{
\dontrun{
if(interactive()){
data(plant_richness_df)
 out <- vif(plant_richness_df[, 5:21])
 out
 }
}
}
