% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cor.R
\name{auto_cor}
\alias{auto_cor}
\title{Multicollinearity reduction via Pearson correlation}
\usage{
auto_cor(
  data = NULL,
  predictor.variable.names = NULL,
  preference.order = NULL,
  max.cor = 0.75,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{predictor.variable.names}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data are used, and non-numeric ones discarded. Default: \code{NULL}}

\item{preference.order}{(optional; character vector) Character vector indicating the user's order of preference to keep variables. Predictors not included in this argument are ranked by the sum of their correlation with other variables (variables with higher sums receive lower ranks and have therefore lower preference order). Default: \code{NULL}.}

\item{max.cor}{(optional; numeric) Numeric between 0 and 1, with recommended values between 0.5 and 0.9. Maximum Pearson correlation between any pair of the selected variables. Default: \code{0.75}}

\item{verbose}{(optional, logical) Logical. if \code{TRUE}, describes the function operations to the user. Default:: \code{TRUE}}
}
\value{
List with three slots:
\itemize{
\item \code{cor}: correlation matrix of the selected variables.
\item \code{selected.variables}: character vector with the names of the selected variables.
\item \code{selected.variables.df}: data frame with the selected variables.
}
}
\description{
Iterative reduction of multicollinearity in a data frame via Pearson bivariate correlation.

The function \code{auto_cor()} applies a recursive algorithm to remove variables with a Pearson correlation with another variable higher than a given threshold (defined by the argument \code{max.cor}).  When two variables are correlated above this threshold, the one with the highest sum of R-squared with all the other variables is removed.

The function \code{auto_vif()} allows the user to define preference selection order via the argument \code{preference.order}. If \code{preference.order} is \verb{"c("a", "b")}, they are correlated above \code{max.cor}, there there are other variables in \code{data}, and the sum of R-squared with all other variables of \code{"a"} is higher than the sum of \code{"b"}, then \code{"b"} is removed anyway.

The argument \code{preference.order} allows the user to "protect" variables that might be interesting or even required for the given analysis.

If \code{preference.order} is not provided, then the predictors are ranked from lower to higher sum of R-squared with the other preodictors, and removed one by one until the maximum R-squared of the correlation matrix is lower than \code{max.cor}.
}
\details{
Can be chained together with \code{\link[=auto_vif]{auto_vif()}} through pipes, see the examples below.
}
\examples{
if(interactive()){

 #loading example data
 data(
  ecoregions_df,
  ecoregions_predictor_variable_names
  )

 #on a data frame
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names
 )

 #getting the correlation matrix
 out$cor

 #getting the names of the selected variables
 out$selected.variables

 #getting the data frame of selected variables
 out$selected.variables.df

 #with preference order (fiver first in ecoregions_predictor_variable_names)
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names,
   preference.order = ecoregions_predictor_variable_names[1:5],
 )

 #with pipes
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names
 ) \%>\%
 auto_vif()

}
}
\seealso{
\code{\link[=auto_vif]{auto_vif()}}
}
