% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cor.R
\name{auto_cor}
\alias{auto_cor}
\title{Multicollinearity reduction via Pearson correlation}
\usage{
auto_cor(
  data = NULL,
  predictor.variable.names = NULL,
  preference.order = NULL,
  cor.threshold = 0.75,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with predictors. Default: \code{NULL}.}

\item{predictor.variable.names}{Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data are used, and non-numeric ones discarded. Default: \code{NULL}}

\item{preference.order}{Character vector indicating the user's order of preference to keep variables. Predictors not included in this argument are ranked by the sum of their correlation with other variables (variables with higher sums receive lower ranks and have therefore lower preference order). Default: \code{NULL}.}

\item{cor.threshold}{Numeric between 0 and 1, with recommended values between 0.5 and 0.9. Maximum Pearson correlation between any pair of the selected variables. Default: \code{0.75}}

\item{verbose}{Logical. if \code{TRUE}, describes the function operations to the user. Default:: \code{TRUE}}
}
\value{
List with three slots:
\itemize{
\item \code{cor}: correlation matrix of the selected variables.
\item \code{selected.variables}: character vector with the names of the selected variables.
\item \code{selected.variables.df}: data frame with the selected variables.
}
}
\description{
Computes the correlation matrix among a set of predictors, orders the correlation matrix according to a user-defined preference order, and removes variables one by one, taking into account the preference order, until the remaining ones are below a given Pearson correlation threshold. \strong{Warning}: variables in \code{preference.order} not in \code{colnames(x)}, and non-numeric columns are removed silently from \code{x} and \code{preference.order}. The same happens with rows having NA values (\code{\link[=na.omit]{na.omit()}} is applied). The function issues a message if zero-variance columns are found.
}
\details{
Can be chained together with \code{\link[=auto_vif]{auto_vif()}} through pipes, see the examples below.
}
\examples{
if(interactive()){

 #loading example data
 data(
  ecoregions_df,
  ecoregions_predictor_variable_names
  )

 #on a data frame
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names
 )

 #getting the correlation matrix
 out$cor

 #getting the names of the selected variables
 out$selected.variables

 #getting the data frame of selected variables
 out$selected.variables.df

 #with preference order (fiver first in ecoregions_predictor_variable_names)
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names,
   preference.order = ecoregions_predictor_variable_names[1:5],
 )

 #with pipes
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_predictor_variable_names
 ) \%>\%
 auto_vif()

}
}
\seealso{
\code{\link[=auto_vif]{auto_vif()}}
}
