% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cor.R
\name{auto_cor}
\alias{auto_cor}
\title{auto_cor}
\usage{
auto_cor(x, preference.order = NULL, cor.threshold = 0.75)
}
\arguments{
\item{x}{a data frame}

\item{preference.order}{character vector indicating the user's order of preference to keep variables. Default: NULL}

\item{cor.threshold}{Pearson correlation threshold, Default: 0.75}
}
\value{
list with two slots:
\itemize{
\item{cor}{correlation matrix of the selected variables}
\item{selected.variables}{character vector with the names of the selected variables.}
}
}
\description{
Reduces the bivariate correlation in a data frame of variables.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)
 out <- auto_cor(x = plant_richness_df[, 4:21])
 }
}
}
