% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cor.R
\name{auto_cor}
\alias{auto_cor}
\title{auto_cor}
\usage{
auto_cor(x = NULL, preference.order = NULL, cor.threshold = 0.75)
}
\arguments{
\item{x}{A data frame with predictors. Default: \code{NULL}.}

\item{preference.order}{Character vector indicating the user's order of preference to keep variables. Default: \code{NULL}.}

\item{cor.threshold}{Numeric between 0 and 1, maximum Pearson correlation between any pair of the selected variables. Default: 0.75}
}
\value{
List with two slots:
\itemize{
\item \strong{cor}: Correlation matrix of the selected variables.
\item \strong{selected.variables}: Character vector with the names of the selected variables.
}
}
\description{
Reduces the bivariate correlation in a data frame of predictors
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)
 out <- auto_cor(x = plant_richness_df[, 4:21])
 out$selected.variables
 }
}
}
