% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_vif.R
\name{auto_vif}
\alias{auto_vif}
\alias{vif}
\title{Multicollinearity reduction via Variance Inflation Factor}
\usage{
auto_vif(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  preference.order = NULL,
  max.vif = 5,
  verbose = TRUE
)

vif(data)
}
\arguments{
\item{data}{(required; data.frame or tibble) A data frame or tibble, or the result of \code{\link[=auto_cor]{auto_cor()}}. Default: \code{NULL}.}

\item{dependent.variable.name}{(optional; character string) Name of the dependent variable. Only required when there are categorical variables within \code{predictor.variable.names}. Default: \code{NULL}}

\item{predictor.variable.names}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data except \code{dependent.variable.name} are used. Default: \code{NULL}}

\item{preference.order}{(optional, character vector) Character vector indicating the preference order to protect variables from elimination.  Predictors not included in this argument are ranked by their VIFs. Default: \code{NULL}.}

\item{max.vif}{(optional, numeric) Numeric between 2.5 and 10 defining the maximum VIF allowed in the output dataset. Higher VIF thresholds should result in a higher number of selected variables. Default: 5.}

\item{verbose}{(optional, logical) Logical. if \code{TRUE}, \code{auto_vif()} prints messages describing its operations on the input data. Default:: \code{TRUE}}
}
\value{
List with three slots:
\itemize{
\item \code{vif}: data frame with the names of the selected variables and their respective VIF scores.
\item \code{selected.variables}: character vector with the names of the selected variables.
\item \code{selected.variables.df}: data frame with the selected variables.
}
}
\description{
NOTE: It is highly recommended to run \code{\link[=auto_cor]{auto_cor()}} before \code{\link[=auto_vif]{auto_vif()}}.

The Variance Inflation Factor for a given variable \code{y} is computed as \code{1/(1-R2)}, where \code{R2} is the multiple R-squared of a multiple regression model fitted using \code{y} as response and all the remaining variables of the input data set as predictors. The equation can be interpreted as "the rate of perfect model's R-squared to the unexplained variance of this model".

The possible range of VIF values is (1, Inf]. A VIF lower than 10 suggest that removing \code{y} from the data set would reduce overall multicollinearity. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5, 5, and 10.

The function \code{auto_vif()} applies a recursive algorithm to remove variables with a VIF higher than a given threshold (defined by the argument \code{max.vif}). However, \code{auto_vif()} allows the user to define preference selection order via the argument \code{preference.order}.

The argument \code{preference.order} allows the user to "protect" variables that might be interesting or even required for the given analysis.

If \code{preference.order} is, for example, \code{c("a", "b", "c")}, \code{max.vif} equals 5, and the VIFs of these variables are 15, 10, and 5, then variable \code{"b"} is removed instead of \code{"a"}.

If \code{preference.order} is not provided, then the predictors are ranked from lower to higher VIF, and removed one by one until their VIF is lower than \code{max.vif}.

If there are categorical variables named in \code{predictor.variable.names} and \code{dependent.variable.name} is provided, then the function applies \code{\link[=fe_target_encoding]{fe_target_encoding()}} with the method "mean" to transform the categorical variables into numeric. If a categorical variable is selected, then its original categorical values are returned.
}
\examples{
if(interactive()){

data(
  ecoregions_df,
  ecoregions_numeric_predictors
  )

#on a data frame
out <- auto_vif(
 data = ecoregions_df,
 predictor.variable.names = ecoregions_numeric_predictors
 )

#getting out the vif data frame
out$vif

#getting the names of the selected variables
out$selected.variables

#getting the data frame of selected variables
out$selected.variables.df

#with preference order
 out <- auto_vif(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_numeric_predictors,
   preference.order = ecoregions_numeric_predictors[1:5],
 )

#with pipes
 out <- auto_cor(
   data = ecoregions_df,
   predictor.variable.names = ecoregions_numeric_predictors
 ) \%>\%
 auto_vif()

 #vif function
 vif(data = ecoregions_df[, ecoregions_numeric_predictors])


}
}
\seealso{
\code{\link[=auto_cor]{auto_cor()}}
}
