% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_vif.R
\name{auto_vif}
\alias{auto_vif}
\title{auto_vif}
\usage{
auto_vif(x, preference.order = NULL, vif.threshold = 5, verbose = FALSE)
}
\arguments{
\item{x}{a data frame}

\item{preference.order}{a character vector with columns names of x ordered by the user preference, Default: NULL}

\item{vif.threshold}{numeric between 5 and 10 defining the selection threshold for the VIF analysis. Default: 5.}

\item{verbose}{when TRUE, describes the function operations to the user.}
}
\value{
list with two slots:
\itemize{
\item{vif}{data frame with the names of the selected variables and their respective VIF scores}
\item{selected.variables}{character vector with the names of the selected variables}
}
}
\description{
Selects variables within a training dataframe that are not linear combinations of other variables by using the variance inflation factor (VIF). This function has two modes of operation:
\itemize{
\item 1. When the argument \code{preference.order} is \code{NULL}, the function removes on on each iteration the variable with the highest VIF until all VIF values are lower than 10. This operation is performed by the hidden function \code{.select_by_max_vif}.
\item 2. When \code{preference.order} is provided, the variables are selected by giving them priority according to their order in \code{preference.order}. If there are variables not in \code{preference.order}, these are selected as in option 1, and finally, all variables are put together and selected by giving priority to the ones in \code{preference.order}. This method preserves the variables desired by the user as much as possible.
}
}
\examples{
\dontrun{
if(interactive()){
 out <- auto_vif(x = plant_richness_df[, 5:20])
 }
}
}
