% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_specification.R
\name{cluster_specification}
\alias{cluster_specification}
\title{cluster_specification}
\usage{
cluster_specification(
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]]
)
}
\arguments{
\item{cluster.ips}{Character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed. Default: \code{NULL}.}

\item{cluster.cores}{Numeric integer vector, number of cores on each machine. Default: \code{NULL}.}

\item{cluster.user}{Character string, name of the user (should be the same throughout machines), Defaults to the current system user.}
}
\value{
A list ready to be used as input for the \emph{spec} argument of the function \link[parallel]{makeCluster}.
}
\description{
Defines a cluster for \link[parallel]{makeCluster}.
}
\details{
This function is used internally by several other functions of the package, but can be useful for a user working with \link[foreach]{foreach} to parallelize loops in a cluster.
}
\examples{
 cluster.spec <- cluster_specification(
   cluster.ips = c("10.0.1.40", "10.0.1.41"),
   cluster.cores = c(6, 4),
   cluster.user = Sys.info()[["user"]]
 )
 cluster.spec
}
