% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_distance_thresholds.R
\name{default_distance_thresholds}
\alias{default_distance_thresholds}
\title{Default distance thresholds to generate spatial predictors}
\usage{
default_distance_thresholds(distance.matrix = NULL)
}
\arguments{
\item{distance.matrix}{Distance matrix. Default: \code{NULL}.}
}
\value{
A numeric vector with distance thresholds.
}
\description{
Generates four distance thresholds, from 0 to max(distance.matrix)/2.
}
\examples{

 #loading example distance matrix
 data(distance_matrix)

 #computing set of default distance thresholds
 default_distance_thresholds(distance_matrix)

}
