% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\docType{data}
\name{distance_matrix}
\alias{distance_matrix}
\title{Matrix of distances among ecoregion edges.}
\format{
A numeric matrix with 227 rows and columns.
}
\usage{
data(distance_matrix)
}
\description{
Distance matrix (in km) among the edges of the American ecoregions described in the \link{plant_richness_df} dataset.
}
\seealso{
\link{plant_richness_df}
}
\keyword{datasets}
