% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_matrix.R
\docType{data}
\name{distance_matrix}
\alias{distance_matrix}
\title{Matrix of distances among american ecoregions.}
\format{
A matrix with 227 rows and columns.
}
\usage{
distance_matrix
}
\description{
Distance matrix in km among the ecoregions described in the \link{plant_richness_df} and \link{plant_richness_sf} dataset.
distance_matrix
}
\keyword{datasets}
