% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_importance.R
\name{get_importance}
\alias{get_importance}
\title{get_importance}
\usage{
get_importance(x)
}
\arguments{
\item{x}{A model produced by \link{rf}, \link{rf_repeat}, or \link{rf_spatial}, or a data frame with importance scores. Default: NULL}
}
\value{
a data frame
}
\description{
Gets variable importance scores from \link{rf}, \link{rf_repeat}, and \link{rf_spatial} models.
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)
data(distance.matrix)
rf.model <- rf(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  distance.matrix = distance_matrix,
  distance.thresholds = c(0, 1000, 2000),
  verbose = FALSE
)
x <- get_importance(x = rf.model)
 }
}
}
