% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moran.R
\name{get_moran}
\alias{get_moran}
\title{get_moran}
\usage{
get_moran(x)
}
\arguments{
\item{x}{A model produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}}
}
\value{
A data frame with moran test results produced by \code{\link[=moran_multithreshold]{moran_multithreshold()}}.
}
\description{
Returns the Moran's I test of a model produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}.
}
\examples{
 data(plant_richness_df)
 data(distance_matrix)

 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000),
   verbose = FALSE
 )

 x <- get_moran(rf.model)
}
