% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Gets out-of-bag performance scores from a model}
\usage{
get_performance(x)
}
\arguments{
\item{x}{Model fitted with \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}.}
}
\value{
A data frame with four columns:
\itemize{
\item \code{metric} Name of the performance metric.
\item \code{mean} Value of the performance metric. Truly a mean only if the model is fitted with \code{\link[=rf_repeat]{rf_repeat()}}.
\item \code{standard_error} Standard error of the mean, only if the model is fitted with \code{\link[=rf_repeat]{rf_repeat()}}, and \code{NA} otherwise.
\item \code{standard_deviation} Standard deviation of the mean, only if the model is fitted with \code{\link[=rf_repeat]{rf_repeat()}}, and \code{NA} otherwise.
}
}
\description{
Returns the performance slot of an \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}} model computed on the out-of-bag data.
}
\examples{
\dontrun{
if(interactive()){

 data(plant_richness_df)
 data(distance.matrix)

 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000),
   verbose = FALSE
 )

 x <- get_performance(rf.model)
 x

}
}
}
\seealso{
\code{\link[=print_performance]{print_performance()}}
}
