% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals.R
\name{get_residuals}
\alias{get_residuals}
\title{get_residuals}
\usage{
get_residuals(x)
}
\arguments{
\item{x}{A model produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}. Default: NULL}
}
\value{
A data frame with residuals, or mean and standard deviation of the residuals if the model was fitted with \code{\link[=rf_repeat]{rf_repeat()}}.
}
\description{
Returns model residuals
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)

rf.model <- rf(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  verbose = FALSE
)

x <- get_residuals(x = rf.model)
 }
}
}
