% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_vif.R
\name{mc_vif}
\alias{mc_vif}
\title{Variance Inflation Factor}
\usage{
mc_vif(data = NULL, predictors.names = NULL, response.name = NULL)
}
\arguments{
\item{data}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{predictors.names}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. If \code{NULL}, all the columns in data except \code{response.name} are used. Default: \code{NULL}}

\item{response.name}{(optional; character string) Name of the dependent variable. Only required when there are categorical variables within \code{predictors.names}. If not provided, non-numeric columns are ignored. Default: \code{NULL}}
}
\value{
Data frame with the columns "variable" containing the predictor name, and "vif" containing the variance inflation factor of the given variable.
}
\description{
Computes the Variance Inflation Factor of all variables in a training data frame.

The Variance Inflation Factor for a given variable \code{y} is computed as \code{1/(1-R2)}, where \code{R2} is the multiple R-squared of a multiple regression model fitted using \code{y} as response and all the remaining variables of the input data set as predictors. The equation can be interpreted as "the rate of perfect model's R-squared to the unexplained variance of this model".

The possible range of VIF values is (1, Inf]. A VIF lower than 10 suggest that removing \code{y} from the data set would reduce overall multicollinearity.

If there are categorical variables named in \code{predictors.names} and \code{response.name} is provided, then the function applies \code{\link[=fe_target_encoding]{fe_target_encoding()}} with the method "mean" to transform the categorical variables into numeric. If a categorical variable is selected, then its original categorical values are returned.

Please note that near-zero variance columns are ignored by this function. Use \code{\link[=mc_auto_vif]{mc_auto_vif()}} to remove them.
}
\examples{
if(interactive()){

data(
  ecoregions_df,
  ecoregions_continuous_response,
  ecoregions_all_predictors
)

df <- mc_vif(
      data = ecoregions_df,
      predictors.names = ecoregions_all_predictors,
      response.name = ecoregions_continuous_response
)

}
}
