% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{mem}
\alias{mem}
\title{mem}
\usage{
mem(x = NULL, distance.threshold = 0, colnames.prefix = "mem")
}
\arguments{
\item{x}{Numeric distance matrix or data frame. Default: \code{NULL}.}

\item{distance.threshold}{Numeric vector with distance thresholds defining different neighborhood extents in the distance matrix, Default: 0}

\item{colnames.prefix}{Character, name prefix for the output columns. Default: \code{"mem"}}
}
\value{
A data frame with positive Moran's Eigenvector Maps.
}
\description{
Computes the positive Moran's Eigenvector Maps of a distance matrix.
}
\details{
The function takes the distance matrix \emph{x}, computes its weights and double-centers it with \code{\link[=double_center_distance_matrix]{double_center_distance_matrix()}}, applies \link[base]{eigen}, and returns eigenvectors with positive normalized eigenvalues.
}
\examples{
\dontrun{
if(interactive()){
 data(distance_matrix)
 distance_matrix.mem <- mem(x = distance_matrix)
 distance_matrix.mem
 }
}
}
