% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem.R
\name{mem}
\alias{mem}
\title{mem}
\usage{
mem(x = NULL, distance.threshold = 0, colnames.prefix = "mem")
}
\arguments{
\item{x}{Numeric matrix or data frame, Default: NULL}

\item{distance.threshold}{distance.thresholds numeric vector with distance thresholds defining neighborhood in the distance matrix, Default: 0}

\item{colnames.prefix}{character, name prefix for the output columns. Default: 'mem'}
}
\value{
A data frame with positive Moran's Eigenvector Maps
}
\description{
Computes the positive Moran's Eigenvector Maps of a distance matrix.
}
\details{
The function takes the distance matrix \code{x}, computes its weights and double-centers it with \link{double_center_distance_matrix}, applies \link[base]{eigen}, and returns eigenvectors with positive normalized eigenvalues (eigenvalues/max(abs(eigenvalues)) > 0).
}
\examples{
\dontrun{
if(interactive()){
 data(distance_matrix)
 distance_matrix.mem <- mem(x = distance_matrix)
 distance_matrix.mem
 }
}
}
