% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\title{Moran's I test on a numeric vector}
\usage{
moran(x = NULL, distance.matrix = NULL, distance.threshold = 0)
}
\arguments{
\item{x}{Numeric vector, generally model residuals, Default: \code{NULL}}

\item{distance.matrix}{Distance matrix among cases in \code{x}. The number of rows of this matrix must be equal to the length of \code{x}. Default: \code{NULL}}

\item{distance.threshold}{numeric value in the range of values available in \code{distance.matrix}. Distances below such threshold are set to 0. Default: \code{0}}
}
\value{
A list with three named slots:
\itemize{
\item \code{moran.i}: Moran's I of \code{x}.
\item \code{p.value}: P-value of the Moran's I estimate.
\item \code{interpretation}: Interpretation of the Moran's I value according to the p-value. One of "Positive spatial correlation", "Negative spatial correlation", and "No spatial correlation".
}
}
\description{
Computes the spatial correlation coefficient (Moran's I) of a vector given a distance matrix, and a distance threshold used to define neihgborhood.
}
\examples{
\dontrun{
if(interactive()){

 data(distance_matrix)
 data(plant_richness)

 moran.out <- moran(
   x = plant_richness$richness_species_vascular,
   distance.matrix = distance_matrix
   )
 moran.out

}
}
}
\seealso{
\code{\link[=moran_multithreshold]{moran_multithreshold()}}
}
