% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_multithreshold.R
\name{pca_multithreshold}
\alias{pca_multithreshold}
\title{pca_multithreshold}
\usage{
pca_multithreshold(
  x = NULL,
  distance.thresholds = 0,
  max.spatial.predictors = 1000
)
}
\arguments{
\item{x}{(required) numeric squared matrix with distances among records, Default: NULL}

\item{distance.thresholds}{(optional) numeric vector with distance thresholds defining neighborhood in the distance matrix, Default: 0}

\item{max.spatial.predictors}{(optional) maximum number of spatial predictors to generate. Only useful when the distance matrix \code{x} is very large. Default: 1000}
}
\value{
a data frame with the PCA factors of the thresholded matrix. The number of columns should be equal to ncol(x) * length(distance.thresholds), but notice that the function \link{pca} removes columns with variance 0, and therefore the final number of columns of the function output may vary. The data frame columns are named "spatial_predictor_DISTANCE_COLUMN", where distance is the given distance threshold, and column is the column index of the given predictor.
}
\description{
computes PCA factors of a distance matrix given a set of distance thresholds. This function allows to generate spatial predictors (Moran's Eigenvector Maps) from the same distance matrix thresholded at different neighborhood distances.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 x <- pca_multithreshold(
   x = distance_matrix,
   distance.thresholds = c(0, 1000)
   )
 x
 colnames(x)
 }
}
}
