% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plants_distance}
\alias{plants_distance}
\title{Distance matrix between ecoregion edges}
\format{
Numeric matrix with 227 rows and 227 columns.
}
\usage{
data(plants_distance)
}
\description{
Distance matrix (in km) between the centroids of American ecoregions in \link{plants_df} computed with \code{\link[=distance_matrix]{distance_matrix()}} and converting the output to \code{km}.
}
\seealso{
\code{\link[=distance_matrix]{distance_matrix()}} for creating custom distance matrices from coordinates.

Other data: 
\code{\link{plants_df}},
\code{\link{plants_predictors}},
\code{\link{plants_response}},
\code{\link{plants_rf}},
\code{\link{plants_rf_spatial}},
\code{\link{plants_xy}}
}
\concept{data}
\keyword{datasets}
