% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evaluation.R
\name{plot_evaluation}
\alias{plot_evaluation}
\title{plot_evaluation}
\usage{
plot_evaluation(x, verbose = TRUE)
}
\arguments{
\item{x}{A model resulting from \code{\link[=rf_evaluate]{rf_evaluate()}}}

\item{verbose}{logical, if TRUE the plot is printed, Default: TRUE}
}
\value{
A ggplot.
}
\description{
Plots the results of an evaluation performed with \code{\link[=rf_evaluate]{rf_evaluate()}}.
}
