% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plots the variable importance of a model}
\usage{
plot_importance(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A model fitted with \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}. Default: \code{NULL}}

\item{verbose}{Logical, if \code{TRUE}, the plot is printed. Default: \code{TRUE}}
}
\value{
A ggplot.
}
\description{
Plots variable importance scores of \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, and \code{\link[=rf_spatial]{rf_spatial()}} models.
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)
data(distance.matrix)

rf.model <- rf(
  data = plant_richness_df,
 dependent.variable.name = "richness_species_vascular",
 predictor.variable.names = colnames(plant_richness_df)[5:21],
 distance.matrix = distance_matrix,
 distance.thresholds = c(0, 1000, 2000),
 verbose = FALSE
)

plot_importance(x = rf.model)

}
}
}
\seealso{
\code{\link[=print_importance]{print_importance()}}, \code{\link[=get_importance]{get_importance()}}
}
