% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{plot_importance}
\usage{
plot_importance(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A model produced by \link{rf}, \link{rf_repeat}, or \link{rf_spatial}, or a data frame with importance scores. Default: NULL}

\item{verbose}{Logical, if TRUE, the generated plot is printed. Default: TRUE}
}
\value{
a ggplot
}
\description{
plots variable importance scores of \link{rf}, \link{rf_repeat}, and \link{rf_spatial} models.
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)
data(distance.matrix)
rf.model <- rf(
  data = plant_richness_df,
 dependent.variable.name = "richness_species_vascular",
 predictor.variable.names = colnames(plant_richness_df)[5:21],
 distance.matrix = distance_matrix,
 distance.thresholds = c(0, 1000, 2000),
 verbose = FALSE
)
plot_importance(x = rf.model)
 }
}
}
