% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization}
\alias{plot_optimization}
\title{Optimization plot of a selection of spatial predictors}
\usage{
plot_optimization(
  model,
  verbose = TRUE
)
}
\arguments{
\item{model}{A model produced by \code{\link[=rf_spatial]{rf_spatial()}}, or an optimization data frame produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}} or \code{\link[=select_spatial_predictors_recursive]{select_spatial_predictors_recursive()}}.}

\item{verbose}{Logical, if \code{TRUE} the plot is printed. Default: \code{TRUE}}
}
\value{
A ggplot.
}
\description{
Plots optimization data frames produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}}
and \code{\link[=select_spatial_predictors_recursive]{select_spatial_predictors_recursive()}}.
}
\details{
If the method used to fit a model with \code{\link[=rf_spatial]{rf_spatial()}} is "hengl", the function returns nothing, as this method does not require optimization.
}
