% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization}
\alias{plot_optimization}
\title{plot_optimization}
\usage{
plot_optimization(x, verbose = TRUE)
}
\arguments{
\item{x}{A model produced by \code{\link[=rf_spatial]{rf_spatial()}}, or an optimization data frame produced by \link{select_spatial_predictors_sequential} or \link{select_spatial_predictors_optimized}.}

\item{verbose}{logical, if TRUE the plot is printed Default: TRUE}
}
\value{
An optimization ggplot.
}
\description{
plots optimization data frames produced by \link{select_spatial_predictors_sequential} and \link{select_spatial_predictors_optimized}.
}
\details{
If the method used to fit the \code{\link[=rf_spatial]{rf_spatial()}} is "hengl", the function returns nothing, as this method does not require optimization.
}
