% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_optimization.R
\name{plot_optimization}
\alias{plot_optimization}
\title{Optimization plot of a selection of spatial predictors}
\usage{
plot_optimization(x, verbose = TRUE)
}
\arguments{
\item{x}{A model produced by \code{\link[=rf_spatial]{rf_spatial()}}, or an optimization data frame produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}} or \code{\link[=select_spatial_predictors_optimized]{select_spatial_predictors_optimized()}}.}

\item{verbose}{logical, if TRUE the plot is printed Default: TRUE}
}
\value{
A ggplot.
}
\description{
Plots optimization data frames produced by \code{\link[=select_spatial_predictors_sequential]{select_spatial_predictors_sequential()}}
and \code{\link[=select_spatial_predictors_optimized]{select_spatial_predictors_optimized()}}.
}
\details{
If the method used to fit a model with \code{\link[=rf_spatial]{rf_spatial()}} is "hengl", the function returns nothing, as this method does not require optimization.
}
