% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tuning.R
\name{plot_tuning}
\alias{plot_tuning}
\title{Plots a tuning object produced by \code{\link[=rf_tuning]{rf_tuning()}}}
\usage{
plot_tuning(x, verbose = TRUE)
}
\arguments{
\item{x}{A model fitted with \code{\link[=rf_tuning]{rf_tuning()}}. Default: \code{NULL}}

\item{verbose}{Logical, if \code{TRUE}, the plot is printed. Default: \code{TRUE}}
}
\value{
A ggplot.
}
\description{
Plots the tuning of the hyperparameters \code{num.trees}, \code{mtry}, and \code{min.node.size} performed by \code{\link[=rf_tuning]{rf_tuning()}}.
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)

rf.model <- rf_tuning(
 data = plant_richness_df,
 dependent.variable.name = "richness_species_vascular",
 predictor.variable.names = colnames(plant_richness_df)[5:21],
 verbose = FALSE
)

plot_tuning(x = rf.model)

}
}
}
\seealso{
\code{\link[=rf_tuning]{rf_tuning()}}
}
