% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_evaluation.R
\name{print_evaluation}
\alias{print_evaluation}
\title{print_evaluation}
\usage{
print_evaluation(x)
}
\arguments{
\item{x}{A model resulting from \code{\link[=rf_evaluate]{rf_evaluate()}}}
}
\value{
A huxtable printed to the standard output.
}
\description{
Prints the results of an evaluation performed with \code{\link[=rf_evaluate]{rf_evaluate()}}.
}
