% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_evaluation.R
\name{print_evaluation}
\alias{print_evaluation}
\title{Prints cross-validation results}
\usage{
print_evaluation(model)
}
\arguments{
\item{model}{A model resulting from \code{\link[=rf_evaluate]{rf_evaluate()}}.}
}
\value{
A table printed to the standard output.
}
\description{
Prints the results of an spatial cross-validation performed with \code{\link[=rf_evaluate]{rf_evaluate()}}.
}
\examples{

#loading example data
data(plants_df)
data(plants_distance)

#fitting random forest model
rf.model <- rf(
  data = plants_df,
  dependent.variable.name = plants_response,
  predictor.variable.names = plants_predictors,
  distance.matrix = plants_distance,
  distance.thresholds = 0,
  n.cores = 1,
  verbose = FALSE
)

#evaluation with spatial cross-validation
rf.model <- rf_evaluate(
  model = rf.model,
  xy = plants_df[, c("x", "y")],
  n.cores = 1
)

#checking evaluation results
print_evaluation(rf.model)

}
\seealso{
\code{\link[=plot_evaluation]{plot_evaluation()}}, \code{\link[=get_evaluation]{get_evaluation()}}
}
