% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_importance.R
\name{print_importance}
\alias{print_importance}
\title{print_importance}
\usage{
print_importance(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A model produced by \link{rf}, \link{rf_repeat}, or \link{rf_spatial}, or a data frame with importance scores. Default: NULL}

\item{verbose}{Logical, if TRUE, variable importance is returned. Default: TRUE}
}
\value{
a ggplot
}
\description{
prints variable importance scores from \link{rf}, \link{rf_repeat}, and \link{rf_spatial} models.
}
\examples{
\dontrun{
if(interactive()){

data(plant_richness_df)
data(distance.matrix)
rf.model <- rf(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  distance.matrix = distance_matrix,
  distance.thresholds = c(0, 1000, 2000),
  verbose = FALSE
)
print_importance(x = rf.model)
 }
}
}
