% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_moran.R
\name{print_moran}
\alias{print_moran}
\title{print_moran}
\usage{
print_moran(x, caption = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A model produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}, or a data frame resulting from \link{moran_multithreshold}. Default: NULL}

\item{caption}{character, caption to attach to the output, Default: NULL}

\item{verbose}{logical, if TRUE, the resulting table is printed into the console, Default: TRUE}
}
\value{
Prints a table in the console using the \link[huxtable]{huxtable} package.
}
\description{
Prints spatial correlation tests.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness_df)
 data(distance.matrix)

 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000),
   verbose = FALSE
 )

 print_moran(rf.model)
 }
}
}
