% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_vector.R
\name{rescale_vector}
\alias{rescale_vector}
\title{rescale_vector}
\usage{
rescale_vector(x = NULL, new.min = 0, new.max = 100, integer = FALSE)
}
\arguments{
\item{x}{numeric vector, Default: NULL}

\item{new.min}{new minimum value, Default: 0}

\item{new.max}{new maximum value, Default: 100}

\item{integer}{coerce the output to integer, Default: FALSE}
}
\value{
a numeric vector of the same length as x, but with its values rescaled between new.min and new.max.
}
\description{
Helper function to rescale numeric vectors to a new range.
}
\examples{
\dontrun{
if(interactive()){
 out <- rescale_vector(
   x = rnorm(100),
   new.min = 0,
   new.max = 100,
   integer = TRUE
   )
   out
 }
}
}
