% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{rf}
\usage{
rf(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  seed = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  trees.per.variable = NULL,
  scaled.importance = TRUE,
  ranger.arguments = list(formula = NULL, mtry = NULL, importance = "permutation",
    write.forest = TRUE, probability = FALSE, min.node.size = NULL, max.depth = NULL,
    replace = TRUE, case.weights = NULL, class.weights = NULL, splitrule = NULL,
    num.random.splits = 1, alpha = 0.5, minprop = 0.1, split.select.weights = NULL,
    always.split.variables = NULL, respect.unordered.factors = NULL,
    scale.permutation.importance = TRUE, local.importance = FALSE, regularization.factor
    = 1, regularization.usedepth = FALSE, keep.inbag = FALSE,      inbag = NULL, holdout
    = FALSE, quantreg = FALSE, oob.error = TRUE, num.threads = parallel::detectCores() -
    1, save.memory = FALSE, verbose = TRUE, seed = NULL, classification = NULL, x = NULL,
    y = NULL, sample.fraction = 1)
)
}
\arguments{
\item{data}{(required) data frame with a response variable and a set of (preferably uncorrelated) predictors, Default: NULL}

\item{dependent.variable.name}{(required) string with the name of the response variable. Must be in the column names of 'data', Default: NULL}

\item{predictor.variable.names}{(required) character vector with the names of the predictive variables. Every element must be in the column names of 'data', Default: NULL}

\item{seed}{(optional) integer, random seed to facilitate reproducibility, Default: NULL}

\item{distance.matrix}{(optional) a squared matrix with the distances among the records in 'data'. Notice that the rows of 'distance.matrix' and 'data' must be the same. If not provided, the computation of the Moran's I of the residuals is ommited. Default: NULL``}

\item{distance.thresholds}{(optional) numeric vector, distances below each value in the distance matrix are set to 0 for the computation of Moran's I. If NULL, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: NULL.}

\item{trees.per.variable}{(optional) number of individual regression trees to fit per variable in 'predictor.variable.names'. This is an alternative way to define ranger's 'num.trees'. If NULL, 'num.trees' is 500. Default: NULL}

\item{scaled.importance}{(optional) boolean. If TRUE, and 'importance = "permutation', the function scales 'data' with \link{scale_robust} and fits a new model to compute scaled variable importance scores. Default: TRUE}

\item{ranger.arguments}{list with \link[ranger]{ranger} arguments. All \link[ranger]{ranger} arguments are set to their default values except for 'importance', that is set to 'permutation' rather than 'none'. Please, consult the help file of \link[ranger]{ranger} if you are not familiar with the arguments of this function.  Default: list(formula = NULL, mtry = NULL, importance = "permutation",
write.forest = TRUE, probability = FALSE, min.node.size = NULL,
max.depth = NULL, replace = TRUE, case.weights = NULL, class.weights = NULL,
splitrule = NULL, num.random.splits = 1, alpha = 0.5, minprop = 0.1,
split.select.weights = NULL, always.split.variables = NULL,
respect.unordered.factors = NULL, scale.permutation.importance = TRUE,
local.importance = FALSE, regularization.factor = 1, regularization.usedepth = FALSE,
keep.inbag = FALSE, inbag = NULL, holdout = FALSE, quantreg = FALSE,
oob.error = TRUE, num.threads = parallel::detectCores() -
1, save.memory = FALSE, verbose = TRUE, seed = NULL,
classification = NULL, x = NULL, y = NULL, sample.fraction = 1)}
}
\value{
a ranger model with several new slots:
\itemize{
\item{ranger.arguments}{stores the values of the arguments used to fit the ranger model}
\item{variable.importance}{a list containing the vector of variable importance as originally returned by ranger (scaled or not depending on the value of 'scaled.importance'), a data frame with the predictors ordered by their importance, and a ggplot showing the importance values}
\item{pseudo.r.squared}{computed as the correlation between the observations and the predictions}
\item{rmse}{as computed by \link{root_mean_squared_error} with 'normalization = NULL'}
\item{nrmse}{as computed by \link{root_mean_squared_error} with 'normalization = "iq'}
\item{residuals}{computed as observations minus predictions}
\item{spatial.correlation.residuals}{the result of \link{moran_multithreshold}}
}
}
\description{
a convenient wrapper for \link[ranger]{ranger} that completes its output by providing the Moran's I of the residuals for different distance thresholds, the rmse and nrmse (as computed by \link{root_mean_squared_error}), and variable importance scores based on a scaled version of the data generated by \link{scale_robust}.
}
\details{
Please read the help file of \link[ranger]{ranger} for further details.
}
\examples{
\dontrun{
if(interactive()){
 data("plant_richness_df")
 data("distance_matrix")
 out <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 100, 1000, 10000)
 )

 class(out)

 #data frame with ordered variable importance
 out$variable.importance$df

 #variable importance plot
 out$variable.importance$plot

 #pseudo R-squared
 out$pseudo.r.squared

 #rmse and nrmse
 out$rmse
 out$nrmse

 #spatial correlation of the residuals for different distance thresholds
 out$spatial.correlation.residuals$df

 #plot of the Moran's I of the residuals for different distance thresholds
 out$spatial.correlation.residuals$plot
 }
}
}
