% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.R
\name{rf}
\alias{rf}
\title{Random forest models with Moran's I test of the residuals}
\usage{
rf(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  ranger.arguments = NULL,
  scaled.importance = TRUE,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{dependent.variable.name}{Character string with the name of the response variable. Must be in the column names of \code{data}. Default: \code{NULL}}

\item{predictor.variable.names}{Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Optionally, the result of \code{\link[=auto_cor]{auto_cor()}} or \code{\link[=auto_vif]{auto_vif()}} Default: \code{NULL}}

\item{distance.matrix}{Squared matrix with the distances among the records in \code{data}. The number of rows of \code{distance.matrix} and \code{data} must be the same. If not provided, the computation of the Moran's I of the residuals is omitted. Default: \code{NULL}}

\item{distance.thresholds}{Numeric vector with neighborhood distances. All distances in the distance matrix below each value in \code{dustance.thresholds} are set to 0 for the computation of Moran's I. If \code{NULL}, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: \code{NULL}}

\item{ranger.arguments}{Named list with \link[ranger]{ranger} arguments (other arguments of this function can also go here). All \link[ranger]{ranger} arguments are set to their default values except for 'importance', that is set to 'permutation' rather than 'none'. The ranger arguments \code{x}, \code{y}, and \code{formula} are disabled. Please, consult the help file of \link[ranger]{ranger} if you are not familiar with the arguments of this function.}

\item{scaled.importance}{Logical, if \code{TRUE}, the function scales \code{data} with \code{\link[=scale_robust]{scale_robust()}} and fits a new model to compute scaled variable importance scores. This makes variable importance scores of different models somewhat comparable. Default: \code{TRUE}}

\item{seed}{Integer, random seed to facilitate reproducibility. If set to a given number, the returned model is always the same. Default: \code{NULL}}

\item{verbose}{Boolean. If TRUE, messages and plots generated during the execution of the function are displayed. Default: \code{TRUE}}
}
\value{
A ranger model with several new slots:
\itemize{
\item \code{ranger.arguments}: Stores the values of the arguments used to fit the ranger model.
\item \code{variable.importance}: A list containing a data frame with the predictors ordered by their importance, and a ggplot showing the importance values.
\item \code{performance}: out-of-bag performance scores: R squared, pseudo R squared, RMSE, and normalized RMSE (NRMSE).
\item \code{pseudo.r.squared}: computed as the correlation between the observations and the predictions.
\item \code{residuals}: computed as observations minus predictions.
\item \code{spatial.correlation.residuals}: the result of \code{\link[=moran_multithreshold]{moran_multithreshold()}} applied to the model results.
}
}
\description{
A convenient wrapper for \link[ranger]{ranger} that completes its output by providing the Moran's I of the residuals for different distance thresholds, the rmse and nrmse (as computed by \code{\link[=root_mean_squared_error]{root_mean_squared_error()}}), and variable importance scores based on a scaled version of the data generated by \code{\link[=scale_robust]{scale_robust()}}.
}
\details{
Please read the help file of \link[ranger]{ranger} for further details. Notice that the \code{formula} interface of \link[ranger]{ranger} is supported through \code{ranger.arguments}, but variable interactions are not allowed (but check \code{\link[=rf_interactions]{rf_interactions()}}).
}
\examples{
\donttest{
if(interactive()){

 data("plant_richness_df")
 data("distance_matrix")

 out <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 100, 1000, 10000)
 )

 class(out)

 #data frame with ordered variable importance
 out$variable.importance$per.variable

 #variable importance plot
 out$variable.importance$plot

 #pseudo R-squared
 out$pseudo.r.squared

 #rmse and nrmse
 out$rmse
 out$nrmse

 #spatial correlation of the residuals for different distance thresholds
 out$spatial.correlation.residuals$df

 #plot of the Moran's I of the residuals for different distance thresholds
 out$spatial.correlation.residuals$plot

 #predictions for new data as done with ranger models:
 predicted <- stats::predict(
   object = out,
   data = plant_richness_df,
   type = "response"
 )$predictions

 #alternative data input methods
 ###############################

 #ranger.arguments can contain ranger arguments and any other rf argument
 my.arguments <- list()
 my.arguments$data <- plant_richness_df
 my.arguments$dependent.variable.name <- "richness_species_vascular"
 my.arguments$predictor.variable.names <- colnames(plant_richness_df)[8:21]
 my.arguments$distance.matrix <- distance_matrix
 my.arguments$distance.thresholds <- c(0, 100, 1000)
 out <- rf(ranger.arguments = my.arguments)

 #the formula input method of ranger is supported (requires 'data')
 #variable interactions are not supported by this interface
 out <- rf(
   data = plant_richness_df,
   ranger.arguments = list(
     formula = as.formula(
     "richness_species_vascular ~
     climate_hypervolume +
     climate_velocity_lgm_average +
     neighbors_count +
     neighbors_percent_shared_edge"
     )
   )
 )

 #the x and y interface is also supported (the name of the response variable is converted to "y")
 out <- rf(
   ranger.arguments = list(
     y = plant_richness_df$richness_species_vascular,
     x = plant_richness_df[, 5:21]
     )
   )

}
}
}
