% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_compare.R
\name{rf_compare}
\alias{rf_compare}
\title{Compares the performance of several models on independent data}
\usage{
rf_compare(
  models = NULL,
  xy = NULL,
  repetitions = 30,
  training.fraction = 0.8,
  metrics = c("r.squared", "pseudo.r.squared", "rmse", "nrmse"),
  notch = TRUE,
  verbose = TRUE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = 11000
)
}
\arguments{
\item{models}{Named list with models based on the same pairs of coordinates. Example: \code{models = list(a = model.a, b = model.b)}. Default: \code{NULL}}

\item{xy}{Data frame or matrix with two columns containing coordinates and named "x" and "y", or an sf file with geometry class \code{sfc_POINT} (see \link{plant_richness_sf}). The same \code{xy} is used for all models, that's why all models must be based on the same pairs of coordinates. If \code{NULL}, the function will throw an error. Default: \code{NULL}}

\item{repetitions}{Integer, must be lower than the total number of rows available in the model's data. Default: \code{30}}

\item{training.fraction}{Proportion between 0.5 and 0.9 indicating the number of records to be used in model training. Default: \code{0.8}}

\item{metrics}{Character vector, names of the performance metrics selected. The possible values are: "r.squared" (\code{cor(obs, pred) ^ 2}), "pseudo.r.squared" (\code{cor(obs, pred)}), "rmse" (\code{sqrt(sum((obs - pred)^2)/length(obs))}), "nrmse" (\code{rmse/(quantile(obs, 0.75) - quantile(obs, 0.25))}). Default: \code{c("r.squared", "pseudo.r.squared", "rmse", "nrmse")}}

\item{notch}{Logical, if \code{TRUE}, boxplot notches are plotted. Default: \code{TRUE}}

\item{verbose}{Logical. If \code{TRUE}, messages and plots generated during the execution of the function are displayed, Default: \code{TRUE}}

\item{n.cores}{Integer, number of cores to use during computations. If \code{NULL}, all cores but one are used, unless a cluster is used. Default = \code{NULL}}

\item{cluster.ips}{Character vector with the IPs of the machines in a cluster. The machine with the first IP will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{Numeric integer vector, number of cores to use on each machine.}

\item{cluster.user}{Character string, name of the user (should be the same throughout machines). Defaults to the current system user.}

\item{cluster.port}{Integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: \code{11000}}
}
\value{
A list with three slots: \code{df.long}, with the performance metrics for each model across repetitions; \code{df.aggregated}, with aggregated stats per model and performance metrics; \code{plot}, with a boxplot built from \code{df.long}.
}
\description{
Uses \code{\link[=rf_evaluate]{rf_evaluate()}} to compare the performance of several models based on the same pairs of coordinates on independent spatial folds.
}
\examples{
\donttest{
if(interactive()){
 data(distance_matrix)
 data(plant_richness_df)

 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000)
 )

 rf.spatial <- rf_spatial(model = rf.model)

 comparison <- rf_compare(
   a = rf.model,
   b = rf.spatial,
   a.name = "Non spatial",
   b.name = "Spatial",
   xy = plant_richness_df[, c("x", "y")],
   )
 }
}
}
\seealso{
\code{\link[=rf_evaluate]{rf_evaluate()}}
}
