% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_repeat.R
\name{rf_repeat}
\alias{rf_repeat}
\title{rf_repeat}
\usage{
rf_repeat(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  ranger.arguments = NULL,
  trees.per.variable = NULL,
  scaled.importance = TRUE,
  repetitions = 5,
  keep.models = FALSE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = 11000
)
}
\arguments{
\item{data}{(required) data frame with a response variable and a set of (preferably uncorrelated) predictors, Default: NULL}

\item{dependent.variable.name}{(required) string with the name of the response variable. Must be in the column names of 'data', Default: NULL}

\item{predictor.variable.names}{(required) character vector with the names of the predictive variables. Every element must be in the column names of 'data', Default: NULL}

\item{distance.matrix}{(optional) a squared matrix with the distances among the records in 'data'. Notice that the rows of 'distance.matrix' and 'data' must be the same. If not provided, the computation of the Moran's I of the residuals is ommited. Default: NULL.}

\item{distance.thresholds}{(optional) numeric vector, distances below each value in the distance matrix are set to 0 for the computation of Moran's I. If NULL, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: NULL.}

\item{ranger.arguments}{(optional) list with \link[ranger]{ranger} arguments. All \link[ranger]{ranger} arguments are set to their default values except for 'importance', that is set to 'permutation' rather than 'none'. Please, consult the help file of \link[ranger]{ranger} if you are not familiar with the arguments of this function.}

\item{trees.per.variable}{(optional) number of individual regression trees to fit per variable in 'predictor.variable.names'. This is an alternative way to define ranger's 'num.trees'. If NULL, 'num.trees' is 500. Default: NULL}

\item{scaled.importance}{(optional) boolean. If TRUE, and 'importance = "permutation', the function scales 'data' with \link{scale_robust} and fits a new model to compute scaled variable importance scores. Default: TRUE}

\item{repetitions}{integer, number of random forest models to fit. Default: 5}

\item{keep.models}{boolean, if TRUE, the fitted models are returned in the "models" slot. Default: FALSE.}

\item{n.cores}{number of cores to use to compute repetitions. If NULL, all cores but one are used, unless a cluster is used.}

\item{cluster.ips}{character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{numeric integer vector, number of cores on each machine.}

\item{cluster.user}{character string, name of the user (should be the same throughout machines), Defaults to the current system user. Default: user name of the current session.}

\item{cluster.port}{integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: 11000.}
}
\value{
a ranger model with several new slots:
\itemize{
\item{ranger.arguments}{stores the values of the arguments used to fit the ranger model}
\item{predictions}{a list with the predictions obtained on each repetition stored in a data frame named 'df.wide' and the average of the predictions in a data frame named 'df'}
\item{variable.importance}{a list containing a data frame with the variable importance obtained on each iteration (df.wide), the mean importance of each predictor across repetitions (df), a long version of the df.wide data frame to facilitate plotting (df.long), and a boxplot showing the distribution of the importance scores across repetitions}
\item{pseudo.r.squared}{pseudo R-squared values throughout repetitions}
\item{rmse}{rmse obtained on each repetition}
\item{nrmse}{normalizad rmse obtained on each repetition}
\item{residuals}{the residuals obtained on each repetition (df.wide), their mean (df) and their stats (stats)}
\item{spatial.correlation.residuals}{the result of \link{moran_multithreshold} applied to the results of each repetition (df.long), the mean of Moran's I across repetitions (df), and a plot with the results of every repetition (plot)}
}
}
\description{
Repeats a given random forest model several times in order to capture the effect of the stochasticity of the algorithm on importance scores and accuracy measures. The function is prepared to run on a cluster if the IPs, number of cores, and user name are provided (see \link{cluster_specification}).
}
\details{
Please read the help file of \link{rf} and \link[ranger]{ranger} for further details.
}
\examples{
\dontrun{
if(interactive()){
 data("plant_richness_df")
 data("distance_matrix")
 out <- rf_repeat(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 100, 1000, 10000),
   repetitions = 10,
   n.cores = 1
 )

 class(out)

 #data frame with ordered variable importance
 out$variable.importance$df

 #variable importance plot
 out$variable.importance$plot

 #pseudo R-squared
 out$pseudo.r.squared

 #rmse and nrmse
 out$rmse
 out$nrmse

 #spatial correlation of the residuals for different distance thresholds
 out$spatial.correlation.residuals$df

 #plot of the Moran's I of the residuals for different distance thresholds
 out$spatial.correlation.residuals$plot
 }
}
}
