% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_select.R
\name{rf_select}
\alias{rf_select}
\title{Selects uncorrelated predictors for Random Forest models}
\usage{
rf_select(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  ranger.arguments = NULL,
  repetitions = 10,
  vif.threshold = 5,
  cor.threshold = 0.5,
  seed = 1,
  verbose = TRUE,
  n.cores = parallel::detectCores() - 1,
  cluster = NULL
)
}
\arguments{
\item{data}{Data frame with a response variable and a set of predictors. Default: \code{NULL}}

\item{dependent.variable.name}{Character string with the name of the response variable. Must be in the column names of \code{data}. If the dependent variable is binary with values 1 and 0, the argument \code{case.weights} of \code{ranger} is populated by the function \code{\link[=case_weights]{case_weights()}}. Default: \code{NULL}}

\item{predictor.variable.names}{Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{ranger.arguments}{Named list with \link[ranger]{ranger} arguments (other arguments of this function can also go here). All \link[ranger]{ranger} arguments are set to their default values except for 'importance', that is set to 'permutation' rather than 'none'. Please, consult the help file of \link[ranger]{ranger} if you are not familiar with the arguments of this function.}

\item{repetitions}{Integer, number of random forest models to fit. Default: \code{10}}

\item{vif.threshold}{Numeric between 2.5 and 10 defining the selection threshold for the VIF analysis. Higher numbers result in a more relaxed variable selection. Default: 5.}

\item{cor.threshold}{Numeric between 0 and 1, with recommended values between 0.5 and 0.9. Maximum Pearson correlation between any pair of the selected variables. Default: \code{0.50}}

\item{seed}{Integer, random seed to facilitate reproduciblity. If set to a given number, the results of the function are always the same. Default: \code{1}.}

\item{verbose}{Logical, ff \code{TRUE}, messages and plots generated during the execution of the function are displayed, Default: \code{TRUE}}

\item{n.cores}{Integer, number of cores used by \code{\link[ranger]{ranger}} for parallel execution (used as value for the argument \code{num.threads} in \code{ranger()}). Default: \code{parallel::detectCores() - 1}}

\item{cluster}{A cluster definition generated with \code{parallel::makeCluster()} or \code{\link{make_cluster}}. Only advisable if you need to spread a large number of repetitions over the nodes of a large cluster when working with large data. If provided, overrides \code{n.cores}. The function does not stop a cluster, please remember to shut it down with \code{\link[=stop_cluster]{stop_cluster()}} or \code{parallel::stopCluster(cl = cluster_name)} at the end of your pipeline. Default: \code{NULL}}
}
\value{
List of class "variable_selection" with five slots:
\itemize{
\item \code{univariate.importance}: a data frame with the names of the predictors and their univariate out-of-bag RMSE.
\item \code{cor}: correlation matrix of the selected variables.
\item \code{vif}: data frame with the names of the selected variables and their respective VIF scores.
\item \code{selected.variables}: character vector with the names of the selected variables.
\item \code{selected.variables.df}: data frame with the selected variables.
}
}
\description{
Combines the functions \code{\link[=rf_repeat]{rf_repeat()}}, \code{\link[=auto_cor]{auto_cor()}}, and \code{\link[=auto_vif]{auto_vif()}} to select a set of uncorrelated predictors ranked by their univariate effect on the response variable. For each predictor in \code{predictor.variable.names}, a univariate model is fitted \code{repetitions} times with \code{\link[=rf_repeat]{rf_repeat()}}, and its out-of-bag RMSE is stored. All variables are then ranked by their RMSE (from minimum to maximum), and their rankings are used as \code{preference.order} in the functions \code{\link[=auto_cor]{auto_cor()}}, and \code{\link[=auto_vif]{auto_vif()}}. The output of this function is a list of the class "variable_selection" that can be used as input for the argument \code{predictor.variable.names} in most modelling functions of this package.

Please, take in mind that this experimental function is not intended to help you increase sheer model performance, but to reduce multicollinearity to increase the interpretability of the variable importance scores.
}
\examples{
#loading example data
data(
  ecoregions_df,
  ecoregions_distance_matrix,
  ecoregions_predvar_names,
  ecoregions_depvar_name
)

#variable selection
selected.variables <- rf_select(
  data = ecoregions_df,
  dependent.variable.name = ecoregions_depvar_name,
  predictor.variable.names = ecoregions_predvar_names,
  repetitions = 10,
  cor.threshold = 0.50,
  vif.threshold = 5,
  seed = 1,
  n.cores = 1,
  verbose = TRUE
)

#fitting a random forest model with the selected variables
rf.model <- rf(
  data = ecoregions_df,
  dependent.variable.name = ecoregions_depvar_name,
  predictor.variable.names = selected.variables,
  distance.matrix = ecoregions_distance_matrix,
  distance.thresholds = 0,
  n.cores = 1
)

}
