% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_tuning.R
\name{rf_tuning}
\alias{rf_tuning}
\title{rf_tuning}
\usage{
rf_tuning(
  model = NULL,
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  tuning.method = "spatial.cv",
  num.trees = c(500, 1000),
  trees.per.variable = NULL,
  mtry = c(2, 3),
  min.node.size = c(5, 10),
  xy = NULL,
  repetitions = 30,
  training.fraction = 0.6,
  verbose = TRUE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = NULL,
  cluster.port = 11000
)
}
\arguments{
\item{model}{Model to tune, produced by \code{\link[=rf]{rf()}}, \code{\link[=rf_repeat]{rf_repeat()}}, or \code{\link[=rf_spatial]{rf_spatial()}}.}

\item{data}{data frame with a response variable and a set of (preferably uncorrelated) predictors. Not required if \code{model} is provided. Default: NULL}

\item{dependent.variable.name}{string with the name of the response variable. Must be in the column names of 'data'. Not required if \code{model} is provided. Default: NULL}

\item{predictor.variable.names}{character vector with the names of the predictive variables. Every element must be in the column names of 'data'. Not required if \code{model} is provided. Default: NULL}

\item{tuning.method}{Character, "oob" to use R squared computed on the out-of-bag data to guide the tuning analysis, and "spatial.cv" to use R squared values from a spatial cross-validation done via \code{\link[=rf_evaluate]{rf_evaluate()}}. Default: "spatial.cv".}

\item{num.trees}{Numeric integer vector with the number of trees to fit on each model repetition. Defalut: c(500, 1000).}

\item{trees.per.variable}{(optional) number of individual regression trees to fit per variable in 'predictor.variable.names'. This is an alternative way to define ranger's 'num.trees'. If NULL, 'num.trees' is 500. Default: NULL}

\item{mtry}{Numeric integer vector, number of predictors to randomly select from the complete pool of predictors on each tree split. Default: c(2, 3)}

\item{min.node.size}{Numeric integer, minimal number of cases in a terminal node. Default:  c(5, 10)}

\item{xy}{Data frame or matrix with two columns containing coordinates and named "x" and "y", or an sf file with geometry class \code{sfc_POINT} (see \link{plant_richness_sf}). If \code{NULL}, the function will throw an error. Default: \code{NULL}}

\item{repetitions}{Integer, number of repetitions to compute the R squared from. If \code{tuning.method = "oob"}, number of repetitions to be used in \code{\link[=rf_repeat]{rf_repeat()}} to fit models for each combination of hyperparameters. If \code{tuning.method = "spatial.cv"}, number of independent spatial folds to use during the cross-validation. Default: 10}

\item{training.fraction}{Proportion between 0.2 and 0.8 indicating the number of records to be used in model training. Default: 0.6}

\item{verbose}{Boolean. If TRUE, messages and plots generated during the execution of the function are displayed, Default: TRUE}

\item{n.cores}{number of cores to use to compute repetitions. If NULL, all cores but one are used, unless a cluster is used.}

\item{cluster.ips}{character vector, IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.}

\item{cluster.cores}{numeric integer vector, number of cores on each machine.}

\item{cluster.user}{character string, name of the user (should be the same throughout machines), Defaults to the current system user. Default: user name of the current session.}

\item{cluster.port}{integer, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: 11000.}
}
\value{
A list with four slots: \strong{tuning} data frame with the results of the tuning analysis; \strong{tuning.long}, a long version of the previous data frame; \strong{tuning.plot}, ggplot of \strong{tuning.long}; \strong{ranger.arguments}, a list ready to be used as the argument \code{ranger.arguments} in \code{\link[=rf_repeat]{rf_repeat()}} or \code{\link[=rf_spatial]{rf_spatial()}}.
}
\description{
Tunes the random forest hyperparameters \code{num.trees}, \code{mtry}, and \code{min.node.size}.
}
\details{
The tuning method "oob" uses as reference the R squared computed on the out-of-bag data, while the method "spatial.cv" uses R squared computed on spatially independent data unseen by the model. The R squared values of the latter method will always be lower, but inform better about the capacity of the combinations of hyperparameters to yield more general models.
}
\examples{
\dontrun{
if(interactive()){
data(plant_richness_df)
data(distance_matrix)

rf.model <- rf(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  verbose = FALSE
)

tuning <- rf_tuning(model = rf.model)
 }
}
}
