% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_robust.R
\name{scale_robust}
\alias{scale_robust}
\title{scale_robust}
\usage{
scale_robust(x, center = TRUE, scale = TRUE, preserveScale = FALSE)
}
\arguments{
\item{x}{matrix or data.frame}

\item{center}{substract median (default:TRUE)}

\item{scale}{scale by the median absolute deviation (mad())  (default:FALSE)}

\item{preserveScale}{equalizes scales without changing them (default:FALSE)}
}
\value{
a scaled matrix or dataframe
}
\description{
scales a matrix or a data frame by the median instead of the mean.
}
\details{
Adapted from \link[quantable]{robustscale}.
}
\examples{
data(plant_richness_df)
plant_richness_df.scaled <- scale_robust(x = plant_richness_df[, 5:10])
}
