% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Standard error of the mean of a numeric vector}
\usage{
standard_error(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value.
}
\description{
Computes the standard error of the mean of a numeric vector as \code{round(sqrt(var(x)/length(x)), 3)}
}
\details{
The function removes \code{NA} values before computing the standard error, and rounds the result to 3 decimal places.
}
\examples{
if(interactive()){

 standard_error(runif(10))

}
}
