% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_mode.R
\name{statistical_mode}
\alias{statistical_mode}
\title{Statistical mode of a vector}
\usage{
statistical_mode(x)
}
\arguments{
\item{x}{Numeric or character vector.}
}
\value{
Statistical mode of \code{x}.
}
\description{
Computes the mode of a numeric or character vector
}
\examples{
\donttest{
if(interactive()){

 statistical_mode(c(rep(10, 10), rep(9, 9)))

}
}
}
