% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_target_encoding_methods.R
\name{fe_target_encoding_mean}
\alias{fe_target_encoding_mean}
\alias{fe_target_encoding_rnorm}
\alias{fe_target_encoding_rank}
\alias{fe_target_encoding_loo}
\title{Target-encoding methods}
\usage{
fe_target_encoding_mean(
  data,
  dependent.variable.name,
  categorical.variable.name,
  noise = 0,
  seed = 1,
  verbose = TRUE
)

fe_target_encoding_rnorm(
  data,
  dependent.variable.name,
  categorical.variable.name,
  sd.width = 0.1,
  seed = 1,
  verbose = TRUE
)

fe_target_encoding_rank(
  data,
  dependent.variable.name,
  categorical.variable.name,
  noise = 0,
  seed = 1,
  verbose = TRUE
)

fe_target_encoding_loo(
  data,
  dependent.variable.name,
  categorical.variable.name,
  verbose = TRUE
)
}
\arguments{
\item{data}{(required; data frame, tibble, or sf) A training data frame. Default: \code{NULL}}

\item{dependent.variable.name}{(required; character string) Name of the response. Must be a column name of \code{data}. Default: \code{NULL}}

\item{categorical.variable.name}{(required; character) Name of the categorical variable to encode.}

\item{noise}{(optional; numeric) Numeric with noise values in the range 0-1. Default: \code{0}.}

\item{seed}{(optional; integer) Random seed to facilitate reproducibility. Default: \code{1}}

\item{verbose}{(optional; logical) If TRUE, messages and plots generated during the execution of the function are displayed. Default: \code{TRUE}}

\item{sd.width}{(optional; numeric) Numeric with multiplicator of the standard deviation of each group in the categorical variable, in the range 0.01-1. Default: \code{0.1}}
}
\value{
The input data frame with a target-encoded variable.

#' @examples
if(interactive()){

data(
ecoregions_df,
ecoregions_continuous_response,
ecoregions_all_predictors
)

#the dataframe ecoregions_df contains two categorical variables
unique(ecoregions_df$dominant_landcover)
unique(ecoregions_df$primary_productivity)

#transforming primary_productivity
ecoregions_df <- fe_target_encoding_mean(
data = ecoregions_df,
dependent.variable.name = ecoregions_continuous_response,
categorical.variable.name = "primary_productivity"
)

#the encoded variable is named primary_productivity__encoded_mean
ecoregions_df$primary_productivity__encoded_mean

#correlation with the response
cor(
x = ecoregions_df$plant_richness,
y = ecoregions_df$primary_productivity__encoded_mean
)

#adding noise
ecoregions_df <- fe_target_encoding_mean(
data = ecoregions_df,
dependent.variable.name = ecoregions_continuous_response,
categorical.variable.name = "primary_productivity",
noise = 0.25
)

#the new encoded variable is named primary_productivity__encoded_mean_noise_0.25
ecoregions_df$primary_productivity__encoded_mean_noise_0.25

#correlation with the response
cor(
x = ecoregions_df$plant_richness,
y = ecoregions_df$primary_productivity__encoded_mean_noise_0.25

}
}
\description{
Target-encoding methods
}
