% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{thinning}
\alias{thinning}
\title{thinning}
\usage{
thinning(xy, minimum.distance = NULL)
}
\arguments{
\item{xy}{a data frame with columns named "x" and "y" representing geographic coordinates.}

\item{minimum.distance}{minimum distance among nearby points, in the same units as the coordinates of xy.}
}
\value{
a data frame with the same columns as xy with points separated by the defined minimum distance.
}
\description{
Resamples a set of points with x and y coordinates to impose a minimum distance among nearby points.
}
\details{
Generally to remove redundant points that could produce pseudo-replication.
}
\examples{
\dontrun{
if(interactive()){
 data(plant_richness)
 plant_richness.thin <- thinning(
   x = plant_richness_df,
   minimum.distance = 5 #oints separated by at least 5 degrees
   )
 plant_richness.thin
 }
}
}
