% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_til_n.R
\name{thinning_til_n}
\alias{thinning_til_n}
\title{thinning_til_n}
\usage{
thinning_til_n(xy, n = 30)
}
\arguments{
\item{xy}{a data frame with columns named "x" and "y" representing geographic coordinates.}

\item{n}{number of samples to obtain.}
}
\value{
a data frame with the same columns as xy with a row number close to n.
}
\description{
Resamples a set of points with x and y coordinates by increasing the distance step by step until a given sample size is obtained.
}
\examples{
 data(plant_richness)
 plant_richness.thin <- thinning_til_n(
   x = plant_richness_df,
   n = 100
   )
 plant_richness.thin
}
